### 2017-08-08 #141 @ChrisMuir
Submission: comtradr
API wrapper pkg for the [UN Comtrade Database](https://comtrade.un.org/data/), which features inter-country trade data dating back to the early 1990's. Full API documentation can be found [here](https://comtrade.un.org/data/doc/api/). This package allows users to interact with the API directly from R, and features functions for making queries and importing data.

```
Package: comtradr
Title: Interface with the United Nations Comtrade API
Version: 0.0.2.9000
Authors@R: person("Chris", "Muir", email = "chrismuirRVA@gmail.com", role = c("aut", "cre"))
Description: Interface with and extract data from the United Nations Comtrade API <https://comtrade.un.org/data/>. Comtrade provides country level shipping data for a variety of commodities, these functions allow for easy API query and data returned as a tidy data frame.
        httr,
        utils
NeedsCompilation: no
Author: Chris Muir [aut, cre]
Maintainer: Chris Muir <chrismuirRVA@gmail.com>
Suggests: testthat,
```

https://github.com/ChrisMuir/comtradr

Data retrieval, this is a API package.

Anyone that could benefit from easy access to historical inter-country shipping data. This can include scientists studying supply chains, economists, journalists, etc.

The Comtrade website has a "Getting API Data with R" post ([here](https://comtrade.un.org/data/Doc/api/ex/r)), but there are a number of issues with the code:
1. It mixes up partner countries and reporter countries.
2. It's missing the call to import the reporter countries.
3. It relies on the `rjson` package.
4. No mention of commodity codes, or importing the commodity codes dataset.
5. Nondescript and confusing parameter names (px, ps, r, p, rg, etc.).

I've found a few R packages online, but they're all just the code from the Comtrade tutorial wrapped up in a package. There's a Python API package, [tradedownloader](https://github.com/ejokeeffe/tradedownloader).
