### 2017-08-11 #141 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @ChrisMuir   !

Currently seeking reviewers. It's a good fit and not overlapping.

Thanks for including `goodpractice::gp()` output  - you can leave that to us in the future 😄

---

Reviewers: xxx
Due date: xxx
