### 2017-08-16 #141 @sckott
thanks @AliciaSchep and @rtaph  for reviewing
