### 2017-09-07 #141 @AliciaSchep
Estimated hours spent reviewing: 3.5 hrs

---

Package was easy and fun to use! :tada: Function documentation was thorough and informative.  I liked the example plots and manipulations in the README, although those might be more appropriate for the vignette rather than README.

#### Examples

Examples in package documentation did not run successfully using `devtools::run_examples(run = FALSE)`. It looked like there was some R output included in the documentation -- I think that should be commented out if included. I am guessing the examples were set to not run for purposes of CRAN, but it would be nice to be able to run them locally in an automated way.

#### Package man page

The package should ideally have a man page for the entire package, which might describe briefly what the package does. A good reference for how to make this in roxygen2 is at http://r-pkgs.had.co.nz/man.html#man-packages.

The tests include many expectations each, especially for ct_search. This can make it harder to figure out exactly why the test is failing, because you get the message from the "test_that" call but then have to do some digging to figure out exactly which expectation(s) failed. My understanding is that it is better practice to have more test_that statements, so each one is reporting on a particular behavior. For example, your "ex1" and "ex2" related expectations should probably be in separate tests if those examples are trying to capture different use cases. Also, things like the following block:

  # Throw error with invalid input for param "reporters".
  expect_error(ct_search(reporters = "invalid_reporter",
                         partners = "Germany",
                         countrytable = countrydf,
                         tradedirection = "imports"))
```

should also probably be a separate test block.

The first time I ran the tests, all the expectations for "ex2" failed. Trying again multiple times, it seemed to work. I am not sure what went wrong the first time, so don't really have any suggestion for that one, just figured I'd mention it.

#### Continuous Integration

It would be good to include the Travis Badge on the README, and to additionally test the package on MacOs and Windows.  For Mac, this can be specified in the .travis.yml file.  For Windows, you can use AppVeyor.

### Additional minor suggestions:

These suggestions could affect existing code using package and are also pretty minor so I can understand if you want to ignore!

#### Name of country_lookup and commodity_lookup functions

Most of the package names start with "ct_", except the "country_lookup" and "commodity_lookup" functions. I think it would be nice to go fully with the convention of starting all the functions with the same prefix, and add "ct_" to those two functions as well.

#### Argument ordering for country_lookup function

For the country_lookup function, I think it would be nice for the order of the arguments to be re-arranged so that "type" was optional without having to write-out the "lookuptable" argument.  Currently, if you want just accept the default for type, then you would need to do:

country_lookup(c("Belgium", "vietnam", "brazil"), lookuptable = countrydf)
```

Reordering arguments for type to be last would allow you do:

country_lookup(c("Belgium", "vietnam", "brazil"), countrydf)
```

Having the argument be named "countrytable" for consistency with ct_search might also be nice. Then commodity_lookup could have the argument be "commoditytable".
