### 2017-09-09 #141 @rtaph
- [X] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

    - Mostly there, but no documentation or examples for `api_col_names()`

    - Unit tests exist, but a whole battery of `ct_search` tests fail when you run the full test suite. This is because the API limits the query rate in guest mode and the tests are all run quickly in succession. As a result, the tests all fail. When I run the tests one by one in the console, they pass. They also pass when I pad your test suite with `Sys.sleep(1)` in between calls to the API. More on this below.

    - The package does not use the recommended `snake_case` convention and mixes naming styles (see below). The standard is not hard and fast but it would be nice if a single approach was adopted, perhaps working around the mixed naming conventions in the API.
    - the package does not provide citation details in the README, as per the guideline
    - the package does not yet contain top-level documentation (i.e. typing `?comtrade` will not return a help page)
    - per guidelines, the package ought to add `#' @noRd` to internal functions such as `api_col_names()`

     - TBD :)

---

#### Overall Approach

Let me start by saying that this is a great package and I enjoyed reviewing it. I wish more people in the field used R, and packages like yours make it a little bit easier to do so. I  like that that virtually all the functionality of the original API is available in the package, and your integration of regex into the package philosophy means that you provide a flexible approach to working with an API from the U.N. that is otherwise is a bit stiff (not very human-friendly).

My thoughts below may be thought of as feature upgrades rather than bugs, but I am including them here because some would break the current approach you have adopted. As a result, you could consider implementing them (if you agree with them) before a major release. It's up to you as a package author.

At a high level, my main recommendation would be to simplify the package philosophy to make data access even easier and faster. I think the package could benefit from design changes that a) reduce the number of steps needed to get at the final data, and b) reduce the need to open up help files to remember how to do things or recall how things are named. I found that each time I loaded up this package, I would need to re-read the help pages and do more setup than I cared for in order to get the data I wanted. I think these difficulties can be addressed in a number of ways.

  - **Returning a Data Frame Directly**: From the [httr vignette on writing good R API wrappers](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html), they suggest that API packages should return a helpful object instead of parsed lists. The status code of the API is not the object of interest iself, and so the fact that the data frame is nested within the return object adds a hurdle that is a bit annoying.  This is particularly true in the tidyverse approach where you want consise, tabular inputs and outputs that work in piped anlytical workflow. Although switching to a data frame output would be a breaking change, IMO it is worth it since the end user will want to get at the final data frame.
  - **Converting to Native R Messages and Warnings**: Building on the previous point, I think it would be much nicer to convert the API's validation status codes into errors and warnings that are native to R. This is the recommendation provided by the [httr vignette](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html). You can then use examples to demonstrate how to appoach error handling using the more universal `tryCatch`, `purrr::safely()`, and `purrr::possibly()`. This is a more robust approach because if the disposition of the output is burried in a list that may not be checked/asserted by the end user, you run the risk that you user will not [fail fast](http://adv-r.had.co.nz/Exceptions-Debugging.html#defensive-programming).
  - **Caching Reference Tables**: The fact that one needs to define `countrytable` and `commoditydf` every time the function is called is an added step and I found myself having to open up the help page each time to remember. It would be nice to have this saved as package data and have an updating function (e.g. `ct_update()` that refreshes the data). I think this makes sense because everyone will always be defining these tables-- and since all your users will need to do this in order to use the package, it makes sense to build it in directly into your package as a default. Also, the reference tables do not change much over time, nor do I see their content really change much as a function input. Seeing as they are much more static than dynamic, it makes sense to save them, cache them, and hide them from day-to-day use. For example, you could write a switch to only re-extract the table if the `rawdata$headers$last-modified` header is more recent than the current date and time. This can speed things up and reduce load on the service.
  - **Throttling API requests**: Since the API has user access limits, it would be great if the R package worked around those limits natively by delaying repeat API calls. One way to go about this would be to store the latest timestamp in a purpose-built [package namespace](https://trestletech.com/2013/04/package-wide-variablescache-in-r-package/) (say, via `ct_cache <- new.env()`), and set switches within your functions to call `Sys.sleep()` if they are called to soon.

#### Package Help Page

As mentioned above, it would be useful to add a `?comtrade` and `?comtrade-package` help page. This was my starting point into your package and I couldn't immediately find documentation.

#### Form and Style

Your code was easy to follow and well annotated (thanks!). Some small thoughts:

- I like the `ct_***` convention for the naming of the family of functions-- it helps keep things consistent and is useful for remembering function names when you have autocompletion in an IDE like RStudio. I would recommend all the package functions follow this convention, including `commodity_lookup()` and `country_lookup()`
- I would avoid recyling the package name in example assignments (`comtrade <-`).
- There are different conventions used between names of variable, arguments, and argument parameters. These include `lowercase` (.e.g. "maxrec"), `UPPERCASE` (.e.g. "TOTAL"), `snake_case` (e.g. "return_code"),  `hyphen-case` (.e.g. "re-imports"), `Title Case` (.e.g. "Partner Code"), and `Sentence case` (.e.g. "Unit code"). I realize some of these inconsistencies come from the API itself (and are passed to the API), but it would be nice if the package had the flexibility to accept a single style through and through, as this would remove the burden from the user to need to know (or lookup) the style. Having to always look up the available choices in the help page is a bit of a drag.

The documentation is generally good with plenty of examples. A few topics I think could be added to the documention include:

 - **Oauth Token**: documenting the use of an authorization token.
 - **Regex**: It would be great to state in the documentation that the functions use regular expressions. Also, not all your users may know regex, so pointing them to some help pages on the topic could be great (maybe add to the see-also section of documention).

#### Modularization

The modularization on this package is generally great. The one thing I would split out is the part of the function that constructs the API call from the part that actually calls it. It would be nice to be able to generate the API url for troubleshooting purposes, even if this is not exported, and also save the API url along with the result (especially if caching).

#### User Identification

It is good web etiquette to identify oneself when accessing APIs and scraping the web. This can be done via the [user agent](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html#set-a-user-agent) variable in `httr` and would be good for people using guest mode.

Along the lines of [Munzert et al.]([Automated Data Collection with R](https://www.amazon.com/Automated-Data-Collection-Practical-Scraping/dp/111883481X)), I would set the user agent to something generic like:

``` r
paste(Sys.info()[["user"]], R.version$version.str, version$platform, sep = ", ")
#> [1] "rtaph, R version 3.4.1 (2017-06-30), x86_64-apple-darwin15.6.0"
```

#### Data Format

Is there any utility in allowing the user to specify `fmt`? I recommend removing the flexibility and only letting the API receive JSON, since validation information is only returned for JSON data. I am nervous about differential behaviour and with CSV important metadata is lost. If you do choose to keep this parameter, it would be good to write a whole suite of tests that check that CSV and JSON variants behave the same in every way (especially w.r.t. warnings and errors).

#### API version check

From what I can tell, there is no call to check the API version (is that right?). If there is, it would be nice [assert](https://github.com/hadley/assertthat) the version in the return data so that if the API changes, the user gets a warning about using an older version of `comtradr`.

#### Names
This is really a matter of preference, but I would simplify by removing the option of "human readable" names in the formals, instead providing a dictionary and examples of how to do this outside of the function. This lets the user go back and forth with renaming columns once the data frame is returned.

Another thought is that the "country name" header has a space, and some people may not know how to use backticks to refer to names like this. Again, maybe think about standardizing the naming convention for data frame headers, argument names, and argument parameters.

#### Status 200 Check

Why not have an [asserter](https://github.com/hadley/assertthat) to ensure that your receive `Status: 200`? Maybe this is not needed since you check the validation codes for JSON, but still, it is easy to implement via `rawdata$status_code == 200L` and an added layer of safety.

#### Spelling and Grammar

Tiny mistakes:

- spelling of "ajust" in `comtradr-vignette.Rmd`, line 98.
- spelling of "tomatos" in `comtradr-vignette.Rmd`, lines 90 and 94

#### Unit Tests

Great test coverage! These are especially important for API wrappers since the upstream protocols can change. A few ideas on tests:

 - Why not create separate testthat messages for each expectation, instead of combining them into one big test? "search return values are correct, and fail when expected" is a bit general. This would help debugging and make test failure messages more specific.
 - Is there any reason that `skip_on_cran()` is commented out? This seems to be the sort of package that this function is useful for.

#### Package Data
Why not save `api_col_names` as internal package data instead of a function? Also, why not save it as a named list (with machine readable variants) in order to allow the end user to have a dictionnary to go back and forth between those names?

#### Handling of Metadata
It would be great if the following metadata was stored somewhere along the way, including:

  - **the API timestamp**: When was the data called?
  - **the API URL**: What is the address/definition of the call?
  - **`system.time()` timings**: How long did the operation take to complete?

I think the best would be to store it as an attribute of the return object, or perhaps as a colum in the data frame. You have to be a bit careful here though because some functions outside the tidyverse drop attributes in their return objects. Alternatively, metadata can be stored in a special purpose namespace for internal reference, however this is likley to be _less_ accessible to end users.

#### Other

- There is no date field in DESCRIPTION file. It is not strictly necessary, but it means that running `citation("comtradr")` will not return a proper citation.

- I suggest setting the order of your function's formals in order of importance (or frequency of use). For example, I don't think the api URL is a common call, so moving it down would help for those who like to use [positional matching](http://adv-r.had.co.nz/Functions.html#function-arguments).

- Why not let the `verbose` argument in your functions draw from the existing setting in the options? I.e. `verbose = getOption("verbose", TRUE)`

- The country table is not denormalized (same code if both a reporter and partner). This is nitpicking but I think a nicer approach would be a single row for each country + code, with logical columns that define if they are a reporter and if they are a partner.

- Since I do not think you are using the dots `...` argument for anything, why not add them in to allow users to pass additional parameters to the regex functions (for example, to modify the `ignore.case` or `perl` arguments)?

#### Questions

- Will `fromJSON()` with `simplifyDataFrame = TRUE` always work? Is there a way to test for non-tabular data, or catch errors? What if there is hierarchical nesting?

#### References and Resources

Article and Books:

 - [Best practices for API packages](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html)

 - [Package-Wide Variables/Cache in R Packages](https://trestletech.com/2013/04/package-wide-variablescache-in-r-package/)

 - Munzert, Simon, et al. [Automated Data Collection with R](https://www.amazon.com/Automated-Data-Collection-Practical-Scraping/dp/111883481X). Wiley. 2015.

- [Programming with dplyr](https://rpubs.com/hadley/dplyr-programming)
