### 2017-09-09 #141 @ChrisMuir
@AliciaSchep and @rtaph,

Thank you both so much for the thorough reviews! I've already begun working on suggestions made by @AliciaSchep, I will now start to address all suggestions made by @rtaph.

I'll compile a reply that addresses each point made within the reviews. This may take some time, but I hope to have a reply to you both in 1-2 weeks.

Also, while I'm working on package updates, I have two additional changes that I came up with recently, but have held off as I didn't want to make changes during the review process. Let me know your thoughts on these:

- **Replace `dplyr` import with `magrittr`**
The main use of `dplyr` within my pkg is the pipe (`%>%`), I think it makes sense to avoid importing `dplyr` if possible, due to its size.

- **Import `purrr`**
I've had issues in the past with type safety in function `commodity_lookup` (see issues [#2](https://github.com/ChrisMuir/comtradr/issues/2) and [#3](https://github.com/ChrisMuir/comtradr/issues/3)). I'm currently using `sapply` followed by a series of transformations designed to account for all possible data type returns from `sapply`, given the input parameters....the function is a little convoluted and inelegant. I've rewritten `commodity_lookup` to use `purrr::map` in place of `sapply`, to streamline the function and make it less prone to type safety issues. Only drawback is I'm not sure if this is enough to warrant adding `purrr` to imports, although @rtaph mentioned adding `purrr::safely` and `purrr::possibly`, plus I can also use `purrr::map_char` to replace some calls to `vapply` in other functions, so might be worth it.

Thanks again!
