### 2017-09-09 #141 @rtaph
Hi @ChrisMuir,

Happy to help. If you are only relying on the pipe, then I think your suggestion to import `magrittr` makes sense. Even tidyverse packages such as `purrr` have done this. Note that `purrr` require `magrittr (>= 1.5)`; there may have been some major changes in that version.

As for the second point-- well everyone has their opinion and you could do it in base R with if you [stick away from `sapply()`](http://adv-r.had.co.nz/Functionals.html). I would say that type safety and clarity is a worthwhile tradeoff, especially as your package gains in popularity and use in the open community. `purrr` is a fairly mature package to rely on so I think it's worth it.
