### 2017-09-10 #141 @ChrisMuir
Got it, thanks for sharing your thoughts on those. Definitely leaning towards implementing them both. I'll look into `magrittr` version history, to see if I need to specify a minimum version.

Thanks!
