### 2017-09-12 ropensci/comtradr#4 @ChrisMuir
Pkg Edits Related to rOpenSci Review
I submitted this package to [rOpenSci](https://ropensci.org/), to be considered as part of their collection of open science packages. It is currently under [reivew](https://github.com/ropensci/onboarding/issues/141). As part of the review process, there are a number of changes and package edits to be made. This issue is meant to serve as a central collection of those changes.

### Changes/Updates/Edits

Boxes will be checked when the change has been made and pushed to this repo.

- [x] Create package level docs page.
- [x] Remove `dontrun` tags from the function examples, allow them to be run.
- [x] Move the example plots from the `README` to the `vignette` (this required adding `dplyr` and `ggplot2` to `Suggests` within the `DESCRIPTION` file).
- [x] Split tests up among more `test_that` statements.
- [x] Expand CI. Currently using Travis to test on only Linux (I think only on Linux?)
- [x] For funcs `country_lookup` and `commodity_lookup`, change function names to include prefix "ct_".
- [x] Reorder func args to move required args ahead of optional args.
- [x] Use `snake_case` for all variables throughout the pkg.
- [x] Eliminate spaces and use `snake_case` for all col headers within output data.
- [x] Add citation details to the `README`.
- [x] Add `#' @noRd` to all internal functions.
- [x] Return a data frame, as opposed to a list, from func `ct_search`.
- [x] Convert to Native R Messages and Warnings, in place of nesting api status codes in the output list.
- [x] Cache county and commodity reference tables.
- [x] Throttle API requests.
- [x] Document the use of an authorization token.
- [x] Be explicit about the use of regex (mention it, list resources explaining its use).
- [x] Set user agent for guest users, as a way of identifying the user to the api.
- [x] Eliminate `fmt` arg in func `ct_search`, always return json data from api.
- [x] Fix two typos in vignette ("ajust" and "tomatos").
- [x] Store metadata of each api call (timestamp, url, `system.time()` timings).
- [x] Remove `dplyr` from imports, add `magrittr` to imports.
- [x] Import `purrr`, use `map` functions in place of `sapply` and `vapply`.
- [x] Setup repo on codecov, then add codecov badge to `README`.
- [x] Update `NEWS` to reflect all changes made during the review.
- [x] Submit updated pkg to CRAN.
