### 2017-09-12 da0a50f
tweak docs, remove printed output from examples
