+++ R/country_lookup.R
@@ -23,8 +23,6 @@
 #' # Look up all reporters that contain the terms "korea" and "vietnam"
 #' countrydf <- ct_countries_table()
 #' country_lookup(c("korea", "vietnam"), "reporter", countrydf)
-#' [1] "Dem. People's Rep. of Korea"   "Fmr Dem. Rep. of Vietnam"
-#' [3] "Fmr Rep. of Vietnam"           "Rep. of Korea"
 #' }
 country_lookup <- function(loc, type = c("reporter", "partner"), lookuptable) {
 
