### 2017-09-12 41dc9a2
tweak docs, remove printed output from examples
