### 2017-09-12 00092e4
fix typos, issue #4
