### 2017-09-12 6cc32f4
create pkg level docs, issue #4
