### 2017-09-12 ca69cba
import purrr and magrittr, dont import dplyr, issue #4
