+++ NAMESPACE
@@ -5,4 +5,4 @@ export(country_lookup)
 export(ct_commodities_table)
 export(ct_countries_table)
 export(ct_search)
-importFrom(dplyr,"%>%")
+importFrom(magrittr,"%>%")
