### 2017-09-12 67a238d
replace sapply with purrr::map for type safety, issue #4
