### 2017-09-12 3b83b10
replace vapply with purrr::map_chr, issue #4
