### 2017-09-12 7dd89c2
import %>% from magrittr, not dplyr
