### 2017-09-12 41b4668
bug fix typo in Roxygen docs
