### 2017-09-12 a249920
expand Travis to test on OSX, issue #4
