+++ .travis.yml
@@ -4,6 +4,15 @@ language: R
 sudo: false
 cache: packages
 
+os:
+- linux
+- osx
+
+r:
+- release
+- oldrelease
+- devel
+
 addons:
   apt:
     packages:
