### 2017-09-12 9c58ab3
add appveyor badge to README
