### 2017-09-12 7fa0950
check CI tests with addition of Sys.sleep
