+++ tests/testthat/test-ct_countries_table.R
@@ -6,6 +6,7 @@ test_that("return has correct attributes, and fails when expected", {
   #skip_on_travis()
 
   df <- ct_countries_table()
+  Sys.sleep(3)
 
   # Data type.
   expect_is(df, "data.frame")
