### 2017-09-12 46c602f
add codecov test coverage
