+++ .travis.yml
@@ -5,8 +5,14 @@ sudo: false
 cache: packages
 
 os:
-- linux
-- osx
+  - linux
+  - osx
+
+r_packages:
+  - covr
+
+after_success:
+  - Rscript -e 'library(covr); codecov()'
 
 addons:
   apt:
