+++ README.Rmd
@@ -19,6 +19,7 @@ comtradr
 
 [![Travis-CI Build Status](https://travis-ci.org/ChrisMuir/comtradr.svg?branch=master)](https://travis-ci.org/ChrisMuir/comtradr)
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ChrisMuir/comtradr?branch=master&svg=true)](https://ci.appveyor.com/project/ChrisMuir/comtradr)
+[![codecov](https://codecov.io/github/ChrisMuir/comtradr/branch/master/graphs/badge.svg)](https://codecov.io/github/ChrisMuir/comtradr)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/comtradr)](http://cran.r-project.org/package=comtradr)
 
 R package for interacting with the [UN Comtrade Database](https://comtrade.un.org/data/) public API. UN Comtrade provides historical data on the weights and value of 
