### 2017-09-13 001432e
rename both lookup func to include prefix 'ct_', issue #4
