+++ tests/testthat/test-ct_country_lookup.R
@@ -1,4 +1,4 @@
-context("country_lookup")
+context("ct_country_lookup")
 
 # All tests on the expected return data.
 test_that("lookup return values are correct, and fail when expected", {
@@ -9,17 +9,17 @@ test_that("lookup return values are correct, and fail when expected", {
   Sys.sleep(3)
 
   # Correct return values for reporter countries lookup.
-  expect_equal(country_lookup(loc = c("Korea", "EU"),
-                              type = "reporter",
-                              lookuptable = df),
+  expect_equal(ct_country_lookup(loc = c("Korea", "EU"),
+                                 type = "reporter",
+                                 lookuptable = df),
                c("Dem. People's Rep. of Korea",
                  "EU-28",
                  "Rep. of Korea" ))
 
   # Correct return values for partner countries lookup.
-  expect_equal(country_lookup(loc = c("Korea", "EU"),
-                              type = "partner",
-                              lookuptable = df),
+  expect_equal(ct_country_lookup(loc = c("Korea", "EU"),
+                                 type = "partner",
+                                 lookuptable = df),
                c("Africa CAMEU region, nes",
                  "Dem. People's Rep. of Korea",
                  "Eastern Europe, nes",
@@ -30,23 +30,23 @@ test_that("lookup return values are correct, and fail when expected", {
                  "Rep. of Korea" ))
 
   # Correct return values when input for "loc" not found in lookup table.
-  expect_equal(country_lookup(loc = "not_a_country",
-                              type = "reporter",
-                              lookuptable = df),
+  expect_equal(ct_country_lookup(loc = "not_a_country",
+                                 type = "reporter",
+                                 lookuptable = df),
                "No matching results found")
 
   # Throw error with invalid input for param "loc".
-  expect_error(country_lookup(loc = 533,
-                              type = "reporter",
-                              lookuptable = df))
+  expect_error(ct_country_lookup(loc = 533,
+                                 type = "reporter",
+                                 lookuptable = df))
 
   # Throw error with invalid input for param "type".
-  expect_error(country_lookup(loc = c("Korea", "EU"),
-                              type = "not_reporter",
-                              lookuptable = df))
+  expect_error(ct_country_lookup(loc = c("Korea", "EU"),
+                                 type = "not_reporter",
+                                 lookuptable = df))
 
   # Throw error with invalid input for param "lookuptable".
-  expect_error(country_lookup(loc = c("Korea", "EU"),
-                              type = "reporter",
-                              lookuptable = "lookuptable"))
+  expect_error(ct_country_lookup(loc = c("Korea", "EU"),
+                                 type = "reporter",
+                                 lookuptable = "lookuptable"))
 })
