+++ NAMESPACE
@@ -1,8 +1,8 @@
 # Generated by roxygen2: do not edit by hand
 
-export(commodity_lookup)
-export(country_lookup)
 export(ct_commodities_table)
+export(ct_commodity_lookup)
 export(ct_countries_table)
+export(ct_country_lookup)
 export(ct_search)
 importFrom(magrittr,"%>%")

[…]
