### 2017-09-13 4a981af
update references to lookup funcs
