### 2017-09-13 1ba01fa
move example plots from README to vignette
