+++ README.Rmd
@@ -7,11 +7,7 @@ output:
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
 ```{r, echo = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE, 
-  comment = "#>", 
-  fig.path = "plots/README-"
-)
+knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 ```
 
 comtradr
@@ -54,7 +50,7 @@ devtools::install_github("ChrisMuir/comtradr")
 
 **Example 1**: Return all exports from China to South Korea, United States and Mexico, for all years
 
-```{r, warning = FALSE, message = FALSE}
+```{r}
 library(comtradr)
 
 # First, read in the country code lookup table as a dataframe from Comtrade. 
@@ -77,7 +73,7 @@ str(example1$data)
 
 **Example 2**: Return all exports related to shrimp from Thailand to all other countries, for years 2007 thru 2011
 
-```{r, warning = FALSE, message = FALSE}
+```{r}
 library(comtradr)
 
 # First, read in the commodity code lookup table as a dataframe from Comtrade.
@@ -97,82 +93,4 @@ example2 <- ct_search(reporters = "Thailand",
 
 # Inspect the output
 str(example2$data)
-```
-
-## Visualize
-
-Once the data is collected, we can use it to create some basic visualizations.
-
-**Plot 1**: Using the data collected above in example 1, plot total value (USD) of Chinese exports to Mexico, South Korea and the United States, by year.
-
-```{r plot1, warning = FALSE, message = FALSE}
-# install.packages("ggplot2")
-library(ggplot2)
-
-df <- example1$data
-
-# Create plot.
-ggplot(df, aes(Year, `Trade Value (US$)`, color = factor(Partner), 
-               shape = factor(Partner))) +
-  geom_point(size = 2) +
-  geom_line(size = 1) +
-  scale_x_continuous(limits = c(min(df$Year), max(df$Year)), 
-                     breaks = seq.int(min(df$Year), max(df$Year), 2)) +
-  scale_color_manual(values = c("orange", "blue", "red"), 
-                     name = "Destination\nCountry") +
-  scale_shape_discrete(name = "Destination\nCountry") +
-  labs(title = "Total Value (USD) of Chinese Exports, by Year",
-       x = "year", y = "total value of shipments in USD") +
-  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1))
-```
-
-**Plot 2**: Using the data collected above in example 2, plot the top eight destination countries/areas of Thai shrimp exports, by weight (KG), for 2007 - 2011.
-
-```{r plot2, warning = FALSE, message = FALSE}
-# install.packages("ggplot2")
-# install.packages("dplyr")
-# install.packages("magrittr")
-library(ggplot2)
-library(dplyr)
-
-df <- example2$data
-
-# Create country specific "total weight per year" dataframe for plotting.
-plotdf <- df %>% 
-  group_by_(.dots = c("Partner", "Year")) %>% 
-  summarise(kg = as.numeric(sum(`Netweight (kg)`, na.rm = TRUE))) %>% 
-  as_data_frame()
-
-# Get vector of the top 8 destination countries/areas by total weight shipped 
-# across all years, then subset plotdf to only include observations related 
-# to those countries/areas.
-top8 <- plotdf %>% 
-  group_by(Partner) %>% 
-  summarise(kg = as.numeric(sum(kg, na.rm = TRUE))) %>% 
-  arrange(desc(kg)) %>% 
-  magrittr::extract2("Partner") %>% 
[…]
