### 2017-09-13 0f277c0
add ggplot2 to Suggests, modify size of plots in vignette
