### 2017-09-14 b6f5126
dont build vignettes on both Travis and appveyor
