### 2017-09-15 2b71711
remove vignette args - appveyor skips vignettes by default
