### 2017-09-15 d23c9f8
add rate limit and auth token features
