+++ R/utils.R
@@ -38,3 +38,12 @@ api_col_names <- function() {
     "Flag"
   ))
 }
+
+# Return the current values of env ct_limit_cache as a list.
+get_cache_values <- function() {
+  list(
+    last_query = get("last_query", envir = ct_limit_cache),
+    next_hour_reset = get("next_hour_reset", envir = ct_limit_cache),
+    queries_this_hour = get("queries_this_hour", envir = ct_limit_cache)
+  )
+}
