+++ man/ct_get_reset_time.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rate_limit.R
+\name{ct_get_reset_time}
+\alias{ct_get_reset_time}
+\title{Comtradr rate limit time check}
+\usage{
+ct_get_reset_time()
+}
+\value{
+date and time in which the hourly query limit will reset. Return is
+  an object of class "POSIXct" (see \code{\link{DateTimeClasses}}).
+}
+\description{
+Get the time in which the hourly limit will reset.
+}
+\examples{
+\dontrun{
+ct_get_reset_time()
+
+# Get minutes remaining until limit reset, as numeric value.
+as.double(Sys.time() - ct_get_reset_time())
+}
+}
