+++ NAMESPACE
@@ -4,5 +4,8 @@ export(ct_commodities_table)
 export(ct_commodity_lookup)
 export(ct_countries_table)
 export(ct_country_lookup)
+export(ct_get_remaining_hourly_queries)
+export(ct_get_reset_time)
+export(ct_register_token)
 export(ct_search)
 importFrom(magrittr,"%>%")
