### 2017-09-15 689be85
remove calls to Sys.sleep
