+++ tests/testthat/test-ct_country_lookup.R
@@ -6,7 +6,6 @@ test_that("lookup return values are correct, and fail when expected", {
   #skip_on_travis()
 
   df <- ct_countries_table()
-  Sys.sleep(3)
 
   # Correct return values for reporter countries lookup.
   expect_equal(ct_country_lookup(loc = c("Korea", "EU"),
