+++ tests/testthat/test-ct_commodities_table.R
@@ -6,7 +6,6 @@ test_that("return has correct attributes, and fails when expected", {
   #skip_on_travis()
 
   df <- ct_commodities_table(type = "HS")
-  Sys.sleep(3)
 
   # Data type.
   expect_is(df, "data.frame")
