### 2017-09-15 eedd2a9
add section on rate limits and auth token
