### 2017-09-15 e3bbf8f
fix typo in docs
