### 2017-09-15 10f0b5f
fix typo in docs, remove unused arg from docs
