+++ man/ct_search.Rd
@@ -70,9 +70,6 @@ value is "human".}
 
 \item{codetype}{Trade data classification scheme to use, as a character
 string. See "Details" for a list of a valid inputs.}
-
-\item{token}{Authorization token, as a character string. Default value is
-NULL.}
 }
 \value{
 List of length three, elements are:
@@ -91,7 +88,7 @@ shipping data. Full API docs can be found here:
 }
 \details{
 Basic rate limit restrictions. For details on how to register a
- valid token, see \code{\lnk{ct_register_token}}. For API docs on rate
+ valid token, see \code{\link{ct_register_token}}. For API docs on rate
  limits, see \url{https://comtrade.un.org/data/doc/api/#Limits}
  \itemize{
  \item Without authentication token: 1 request per second, 100 requests
