### 2017-09-17 bc7fd8a
save country & commodity DBs to file, split up tests, issue #4
