### 2017-09-17 c53cdb8
update how 'next_hour_reset' is calculated
