+++ R/rate_limit.R
@@ -9,15 +9,16 @@
 #' ct_get_remaining_hourly_queries()
 #' }
 ct_get_remaining_hourly_queries <- function() {
-  get("queries_this_hour", envir = ct_limit_cache)
+  get("queries_this_hour", envir = ct_env)
 }
+
 
 #' Comtradr rate limit time check
 #'
 #' Get the time in which the hourly limit will reset.
 #'
 #' @return date and time in which the hourly query limit will reset. Return is
-#'   an object of class "POSIXct" (see \code{\link{DateTimeClasses}}).
+#'   a "POSIXct" object (see \code{\link{DateTimeClasses}}).
 #' @export
 #'
 #' @examples \dontrun{
@@ -27,8 +28,14 @@ ct_get_remaining_hourly_queries <- function() {
 #' as.double(Sys.time() - ct_get_reset_time())
 #' }
 ct_get_reset_time <- function() {
-  get("next_hour_reset", envir = ct_limit_cache) + 3600
+  val <- get("next_hour_reset", envir = ct_env)
+  if (!is.null(val)) {
+    val + 3600
+  } else {
+    Sys.time() + 3600
+  }
 }
+
 
 #' Comtradr set API token
 #'
@@ -38,6 +45,7 @@ ct_get_reset_time <- function() {
 #'
 #' @param token char string, valid API token.
 #'
+#' @return Set comtradr API token and update rate limits.
 #' @export
 #'
 #' @examples \dontrun{
@@ -52,6 +60,14 @@ ct_register_token <- function(token) {
   class(ct_options) <- "comtradr_credentials"
   options(comtradr = ct_options)
 
-  # Increase the hourly limit within the env ct_limit_cache.
-  assign("queries_this_hour", 10000, envir = ct_limit_cache)
+  # Change the hourly limit within the env ct_env. Subtract the number of
+  # queries already performed this hour from the update value.
+  queries_this_hour <- get("queries_this_hour", envir = ct_env)
+  curr_account_type <- getOption("comtradr")$comtrade$account_type
+  if (curr_account_type == "standard") {
+    new_hr_limit <- 10000 - (100 - queries_this_hour)
+  } else if (curr_account_type == "premium") {
+    new_hr_limit <- 10000 - (10000 - queries_this_hour)
+  }
+  assign("queries_this_hour", new_hr_limit, envir = ct_env)
 }
