### 2017-09-17 afcedd7
reorder func args, eliminate 'ct_csv_data.R', issue #4
