+++ R/ct_search.R
@@ -10,42 +10,37 @@
 #' @param partners Country(s) that have interacted with the reporter
 #'  country(s), as a character vector. Can either be a vector of country names,
 #'  or "All" to represent all countries.
-#' @param url Base of the Comtrade url string, as a character string.
-#' @param max_rec Max number of records returned from each API call, as an
-#'  integer. If max_rec is set to NULL, then value is determined by whether or
-#'  not an API token has been rergistered. API cap without a token is 50000,
-#'  cap with a valid token is 250000. Default value is NULL. For details on
-#'  how to register a valid token, see \code{\link{ct_register_token}}.
-#' @param type Type of trade, as a character string. Must be either "goods" or
-#'  "services".
+#' @param tradedirection Indication of which trade directions on which to
+#'  focus, as a character vector. Must either be "all", or a vector containing
+#'  any combination of the following: "imports", "exports", "re-imports",
+#'  "re-exports".
 #' @param freq Time frequency of the returned results, as a character string.
 #'  Must be either "annual" or "monthly".
 #' @param startdate Start date of a time period, or "all". Default value is
 #'  "all". If inputing a date, must be string w/ structure "yyyy-mm-dd".
 #' @param enddate End date of a time period, or "all". Default value is "all".
 #'  If inputing a date, must be string w/ structure "yyyy-mm-dd".
-#' @param tradedirection Indication of which trade directions on which to
-#'  focus, as a character vector. Must either be "all", or a vector containing
-#'  any combination of the following: "imports", "exports", "re-imports",
-#'  "re-exports".
 #' @param commodcodes Character vector of commodity codes, or "TOTAL". Valid
 #'  commodity codes as input will restrict the query to only look for trade
 #'  related to those commodities, "TOTAL" as input will return all trade
 #'  between the indicated reporter country(s) and partner country(s). Default
 #'  value is "TOTAL".
-#' @param fmt Indication as to the format of the returned data, as a character
-#'  string. Must be either "json" or "csv". Regardless of the fmt used, the
-#'  return data will be in the form of a tidy dataframe. "json" is the
-#'  suggested value for this parameter, as the API tends to provide more
-#'  detailed feedback on why a query failed when using json.
+#' @param max_rec Max number of records returned from each API call, as an
+#'  integer. If max_rec is set to NULL, then value is determined by whether or
+#'  not an API token has been rergistered. API cap without a token is 50000,
+#'  cap with a valid token is 250000. Default value is NULL. For details on
+#'  how to register a valid token, see \code{\link{ct_register_token}}.
+#' @param type Type of trade, as a character string. Must be either "goods" or
+#'  "services". Default value is "goods".
+#' @param url Base of the Comtrade url string, as a character string. Default
+#'  value is "https://comtrade.un.org/api/get?" and should be changed unless
+#'  the Comtrade changes their endpoint url.
 #' @param colname Should the output dataframe have human-friendly or
 #'  machine-friendly column names. Human-friendly means easy for a human to
 #'  read and understand, and may contain special characters and spaces.
 #'  Machine-friendly means easy for a machine to parse, and may not contain
 #'  special characters or spaces. Must be either "human" or "machine". Default
 #'  value is "human".
-#' @param codetype Trade data classification scheme to use, as a character
-#'  string. See "Details" for a list of a valid inputs.
 #'
 #' @details Basic rate limit restrictions. For details on how to register a
[…]
