+++ man/ct_search.Rd
@@ -4,13 +4,11 @@
 \alias{ct_search}
 \title{Get UN Comtrade data via API}
 \usage{
-ct_search(reporters, partners, url = "https://comtrade.un.org/api/get?",
-  max_rec = NULL, type = c("goods", "services"), freq = c("annual",
-  "monthly"), startdate = "all", enddate = "all",
-  tradedirection = c("all", "imports", "exports", "re-imports", "re-exports"),
-  commodcodes = "TOTAL", fmt = c("json", "csv"), colname = c("human",
-  "machine"), codetype = c("HS", "H0", "H1", "H2", "H3", "H4", "ST", "S1",
-  "S2", "S3", "S4", "BEC", "EB02"))
+ct_search(reporters, partners, tradedirection = c("all", "imports", "exports",
+  "re-imports", "re-exports"), freq = c("annual", "monthly"),
+  startdate = "all", enddate = "all", commodcodes = "TOTAL",
+  max_rec = NULL, type = c("goods", "services"),
+  url = "https://comtrade.un.org/api/get?", colname = c("human", "machine"))
 }
 \arguments{
 \item{reporters}{Country(s) of interest, as a character vector. Can either
@@ -20,16 +18,10 @@ be a vector of country names, or "All" to represent all countries.}
 country(s), as a character vector. Can either be a vector of country names,
 or "All" to represent all countries.}
 
-\item{url}{Base of the Comtrade url string, as a character string.}
-
-\item{max_rec}{Max number of records returned from each API call, as an
-integer. If max_rec is set to NULL, then value is determined by whether or
-not an API token has been rergistered. API cap without a token is 50000,
-cap with a valid token is 250000. Default value is NULL. For details on
-how to register a valid token, see \code{\link{ct_register_token}}.}
-
-\item{type}{Type of trade, as a character string. Must be either "goods" or
-"services".}
+\item{tradedirection}{Indication of which trade directions on which to
+focus, as a character vector. Must either be "all", or a vector containing
+any combination of the following: "imports", "exports", "re-imports",
+"re-exports".}
 
 \item{freq}{Time frequency of the returned results, as a character string.
 Must be either "annual" or "monthly".}
@@ -40,22 +32,24 @@ Must be either "annual" or "monthly".}
 \item{enddate}{End date of a time period, or "all". Default value is "all".
 If inputing a date, must be string w/ structure "yyyy-mm-dd".}
 
-\item{tradedirection}{Indication of which trade directions on which to
-focus, as a character vector. Must either be "all", or a vector containing
-any combination of the following: "imports", "exports", "re-imports",
-"re-exports".}
-
 \item{commodcodes}{Character vector of commodity codes, or "TOTAL". Valid
 commodity codes as input will restrict the query to only look for trade
 related to those commodities, "TOTAL" as input will return all trade
 between the indicated reporter country(s) and partner country(s). Default
 value is "TOTAL".}
 
-\item{fmt}{Indication as to the format of the returned data, as a character
-string. Must be either "json" or "csv". Regardless of the fmt used, the
-return data will be in the form of a tidy dataframe. "json" is the
-suggested value for this parameter, as the API tends to provide more
-detailed feedback on why a query failed when using json.}
+\item{max_rec}{Max number of records returned from each API call, as an
+integer. If max_rec is set to NULL, then value is determined by whether or
+not an API token has been rergistered. API cap without a token is 50000,
+cap with a valid token is 250000. Default value is NULL. For details on
+how to register a valid token, see \code{\link{ct_register_token}}.}
+
+\item{type}{Type of trade, as a character string. Must be either "goods" or
+"services". Default value is "goods".}
+
[…]

[…]
