### 2017-09-17 ae9b9b7
delete call to Sys.sleep
