### 2017-09-17 efeb107
use func 'ct_commodity_db_type'
