### 2017-09-17 af6bfca
use '@noRd' Roxygen tag for internal func, issue #4
