### 2017-09-17 a6a01c4
make all variables snake_case
