+++ R/ct_search.R
@@ -1,6 +1,6 @@
 #' Get UN Comtrade data via API
 #'
-#' Make queries to the UN Comtrade API, data is returned as a tidy dataframe.
+#' Make queries to the UN Comtrade API, data is returned as a tidy data frame.
 #' Comtrade is a DB hosted by the United Nations that houses country-level
 #' shipping data. Full API docs can be found here:
 #' \url{https://comtrade.un.org/data/doc/api/}
@@ -10,17 +10,17 @@
 #' @param partners Country(s) that have interacted with the reporter
 #'  country(s), as a character vector. Can either be a vector of country names,
 #'  or "All" to represent all countries.
-#' @param tradedirection Indication of which trade directions on which to
+#' @param trade_direction Indication of which trade directions on which to
 #'  focus, as a character vector. Must either be "all", or a vector containing
-#'  any combination of the following: "imports", "exports", "re-imports",
-#'  "re-exports".
+#'  any combination of the following: "imports", "exports", "re_imports",
+#'  "re_exports". Default value is "all".
 #' @param freq Time frequency of the returned results, as a character string.
-#'  Must be either "annual" or "monthly".
-#' @param startdate Start date of a time period, or "all". Default value is
+#'  Must be either "annual" or "monthly". Default value is "annual".
+#' @param start_date Start date of a time period, or "all". Default value is
 #'  "all". If inputing a date, must be string w/ structure "yyyy-mm-dd".
-#' @param enddate End date of a time period, or "all". Default value is "all".
+#' @param end_date End date of a time period, or "all". Default value is "all".
 #'  If inputing a date, must be string w/ structure "yyyy-mm-dd".
-#' @param commodcodes Character vector of commodity codes, or "TOTAL". Valid
+#' @param commod_codes Character vector of commodity codes, or "TOTAL". Valid
 #'  commodity codes as input will restrict the query to only look for trade
 #'  related to those commodities, "TOTAL" as input will return all trade
 #'  between the indicated reporter country(s) and partner country(s). Default
@@ -33,14 +33,13 @@
 #' @param type Type of trade, as a character string. Must be either "goods" or
 #'  "services". Default value is "goods".
 #' @param url Base of the Comtrade url string, as a character string. Default
-#'  value is "https://comtrade.un.org/api/get?" and should be changed unless
-#'  the Comtrade changes their endpoint url.
-#' @param colname Should the output dataframe have human-friendly or
+#'  value is "https://comtrade.un.org/api/get?" and should mot be changed
+#'  unless Comtrade changes their endpoint url.
+#' @param col_name Should the output data frame have human-friendly or
 #'  machine-friendly column names. Human-friendly means easy for a human to
-#'  read and understand, and may contain special characters and spaces.
-#'  Machine-friendly means easy for a machine to parse, and may not contain
-#'  special characters or spaces. Must be either "human" or "machine". Default
-#'  value is "human".
+#'  read and understand. Machine-friendly means easy for a machine to parse,
+#'  and may not be intuitive for a human to read and understand. Must be
+#'  either "human" or "machine". Default value is "human".
 #'
 #' @details Basic rate limit restrictions. For details on how to register a
 #'  valid token, see \code{\link{ct_register_token}}. For API docs on rate
@@ -56,13 +55,13 @@
 #'  parameter combinations, they are listed below:
 #'  \itemize{
 #'  \item Between params "reporters", "partners", and the query date range (as
[…]
