### 2017-09-17 ddc4518
col headers snake_case, add is.null to 'get_cache_values', issue #4
