### 2017-09-17 56ce24a
change initial val of 'last_query' to null
