+++ R/zzz.R
@@ -3,7 +3,7 @@
 # 2. Comtrade database of countries.
 # 3. Comtrade database of commodity descriptions and codes.
 ct_env <- new.env()
-assign("last_query", Sys.time(), envir = ct_env)
+assign("last_query", NULL, envir = ct_env)
 assign("next_hour_reset", NULL, envir = ct_env)
 assign("queries_this_hour", 100, envir = ct_env)
 load(
@@ -14,6 +14,7 @@ load(
   system.file("extdata", "commodity_table.rda", package = "comtradr"),
   envir = ct_env
 )
+
 
 # Establish initial credentials for the Comtrade API.
 .onLoad <- function(libname, pkgname) {
