### 2017-09-18 c066882
changes to handle missing DB files
