+++ R/ct_update_databases.R
@@ -10,6 +10,11 @@
 #' it to the \code{comtradr} package directory, and update the DB for use
 #' within the current R session.
 #'
+#' @param force logical, if TRUE, both the country and commodity databases
+#'  will be downloaded, regardless of the status of the DB's on file. Default
+#'  value is FALSE.
+#' @param verbose logical, if TRUE, an update status message will be printed
+#'  to console. Default value is TRUE.
 #' @param commodity_type Trade data classification scheme to use, see
 #'  "details" for a list of the valid inputs. Default value is "HS", which is
 #'  the default "type" of the commodity database on file upon install of
@@ -62,14 +67,16 @@
 #' @examples \dontrun{
 #' ct_update_databases()
 #' }
-ct_update_databases <- function(commodity_type = c("HS", "HS1992", "HS1996",
-                                                "HS2002", "HS2007", "HS2012",
-                                                "SITC", "SITCrev1", "SITCrev2",
-                                                "SITCrev3", "SITCrev4", "BEC",
-                                                "EB02"),
-                             commodity_url = NULL,
-                             reporter_url = NULL,
-                             partner_url = NULL) {
+ct_update_databases <- function(force = FALSE, verbose = TRUE,
+                                commodity_type = c("HS", "HS1992", "HS1996",
+                                                   "HS2002", "HS2007",
+                                                   "HS2012", "SITC",
+                                                   "SITCrev1", "SITCrev2",
+                                                   "SITCrev3", "SITCrev4",
+                                                   "BEC", "EB02"),
+                                commodity_url = NULL,
+                                reporter_url = NULL,
+                                partner_url = NULL) {
 
   # Input validation.
   commodity_type <- match.arg(commodity_type)
@@ -123,14 +130,18 @@ ct_update_databases <- function(commodity_type = c("HS", "HS1992", "HS1996",
   }
 
   # Create output message.
-  msg <- "All DB's are up to date, no action required"
+  if (verbose) {
+    msg <- "All DB's are up to date, no action required"
+  }
 
   # Get the current date/time.
   curr_date <- format(Sys.time(), "%a, %d %b %Y %X %Z")
 
-  # Get the current databases from ct_env.
-  country_df <- get("country_df", envir = ct_env)
-  commodity_df <- get("commodity_df", envir = ct_env)
+  # If "force" is FALSE, get the current databases from ct_env.
+  if (!force) {
+    country_df <- get("country_df", envir = ct_env)
+    commodity_df <- get("commodity_df", envir = ct_env)
+  }
 
   # Get the commodity database from the Comtrade website. Compare the
   # "last-modified" date value within the header to the value in variable
@@ -141,7 +152,8 @@ ct_update_databases <- function(commodity_type = c("HS", "HS1992", "HS1996",
   # Replacement will be for both the current session and within the data dir
   # of the comtradr package.
   res <- httr::GET(commodity_url)
-  if (commodity_type != commodity_df$type[1] ||
+  if (force ||
+      commodity_type != commodity_df$type[1] ||
       httr::headers(res)$`last-modified` > commodity_df$date[1]) {
     # Extract data frame.
     df <- res %>%
@@ -154,16 +166,17 @@ ct_update_databases <- function(commodity_type = c("HS", "HS1992", "HS1996",
     # Save df to data dir of the comtradr package.
     save(
       df,
-      file = system.file("extdata",
-                         "commodity_table.rda",
-                         package = "comtradr"),
+      file = paste0(system.file("extdata", package = "comtradr"),
+                    "/commodity_table.rda"),
[…]
