### 2017-09-18 7a25fec
fix typo, wrap country update in if statement
