### 2017-09-18 99c7600
adding 'get_country_db' and 'get_commodity_db'
