### 2017-09-19 a963fe7
add user agent to all httr::GET calls
