### 2017-09-19 d45f46e
add metadata as attributes to API return object
