### 2017-09-20 2d5db09
add imprtFrom magrittr %>%
