+++ R/ct_search.R
@@ -1,4 +1,4 @@
-#' Get UN Comtrade data via API
+#' Get UN Comtrade data
 #'
 #' Make queries to the UN Comtrade API, data is returned as a tidy data frame.
 #' Comtrade is a DB hosted by the United Nations that houses country-level
@@ -73,14 +73,11 @@
 #'  \item req_duration: total duration of the API call, in seconds.
 #'  }
 #'
-#' @return List of length three, elements are:
-#'  \itemize{
-#'  \item \code{msg}: Brief message on success/failure of the API call.
-#'  \item \code{details}: More detailed message on success/failure of the API
-#'    call.
-#'  \item \code{data}: Data frame object of return data.
-#'  }
+#' @return Data frame of Comtrade shipping data.
+#'
 #' @export
+#'
+#' @importFrom magrittr "%>%"
 #'
 #' @examples \dontrun{
 #' ## Example API call number 1:
@@ -89,7 +86,7 @@
 #' ex_1 <- ct_search(reporters = "China",
 #'                   partners = c("Rep. of Korea", "USA", "Mexico"),
 #'                   trade_direction = "exports")
-#' nrow(ex_1$data)
+#' nrow(ex_1)
 #'
 #' ## Example API call number 2:
 #' # All shipments related to shrimp between Canada and all other countries,
@@ -106,7 +103,7 @@
 #'                   start_date = "2011-01-01",
 #'                   end_date = "2015-01-01",
 #'                   commod_codes = shrimp_codes)
-#' nrow(ex_2$data)
+#' nrow(ex_2)
 #'
 #' # Access metadata
 #' attributes(ex_2)$url
@@ -115,13 +112,31 @@
 #' }
 ct_search <- function(reporters, partners,
                       trade_direction = c("all", "imports", "exports",
-                                          "re-imports", "re-exports"),
+                                          "re_imports", "re_exports"),
                       freq = c("annual", "monthly"),
                       start_date = "all", end_date = "all",
                       commod_codes = "TOTAL", max_rec = NULL,
                       type = c("goods", "services"),
                       url = "https://comtrade.un.org/api/get?",
                       col_name = c("human", "machine")) {
+
+  ## Input validation related to API limits on parameter combinations.
+
+  # Between params "reporters", "partners", and the query date range (as
+  # dictated by the two params "start_date" and "end_date"), only one of
+  # these three may use the catch-all input "All".
+  if (
+    all(
+      any(tolower(reporters) == "all"),
+      any(tolower(partners) == "all"),
+      any(tolower(c(start_date, end_date)) == "all")
+    )
+  ) {
+    stop(paste("between 'reporters', 'partners', and date range, only one",
+               "of these may be 'all'"), call. = FALSE)
+  }
+
+
 
   # Fetch current values within ct_env (these values help manage
   # throttling of API queries).
@@ -190,16 +205,22 @@ ct_search <- function(reporters, partners,
                              as.Date(end_date, format = "%Y-%m-%d"),
                              by = "year") %>%
         as.Date() %>%
-        format(format = "%Y") %>%
-        paste(collapse = ",")
+        format(format = "%Y")
     } else if (freq == "M") {
       date_range <- seq.Date(as.Date(start_date, format = "%Y-%m-%d"),
                              as.Date(end_date, format = "%Y-%m-%d"),
                              by = "month") %>%
         as.Date() %>%
-        format(format = "%Y%m") %>%
-        paste(collapse = ",")
+        format(format = "%Y%m")
     }
+
+    # Check to make sure the total date range is five or fewer months/years.
[…]
