### 2017-09-20 95da68b
no man file for func 'execute_api_request'
