### 2017-09-20 5ee0fe1
update docs to mention use of regex, issue #4
