+++ R/ct_country_lookup.R
@@ -4,33 +4,39 @@
 #' spelling/capitalization. This is a helper function for querying the country
 #' names/spelling used by Comtrade.. It takes as input a vector of
 #' country names, output is any country names that contain any of the input
-#' strings, using regex via the base function grepl (search is case insensitive).
+#' strings, using regex via the base function grepl.
 #' For use with the UN Comtrade API, full API docs can be found at
 #' \url{https://comtrade.un.org/data/doc/api/}
 #'
-#' @param loc Char vector of country names.
+#' @param search_terms Char vector of country names.
 #' @param type The country list to use for the search, valid inputs are
 #'  "reporter" and "partner".
 #'
 #' @return A character vector of country names that are complete or partial
 #'  matches with any of the input country names.
+#'
+#' @details This function uses regular expressions (regex) to find matches
+#'  within the country DB. This means it will treat as a match any country
+#'  string that contains the input search term. For more on using regex
+#'  within R, see this great tutorial by Gloria Li and Jenny Bryan
+#'  \url{http://stat545.com/block022_regular-expression.html}
+#'
 #' @export
+#'
+#' @seealso \code{\link{grepl}}
 #'
 #' @examples
 #' # Look up all reporters that contain the terms "korea" and "vietnam"
 #' ct_country_lookup(c("korea", "vietnam"), "reporter")
-ct_country_lookup <- function(loc, type = c("reporter", "partner")) {
+ct_country_lookup <- function(search_terms, type = c("reporter", "partner")) {
   # Input validation.
   type <- match.arg(type)
-  if (!is.character(loc) || length(loc) < 1) {
-    stop("param 'loc' must be a char vector with length greater than zero",
-         call. = FALSE)
-  }
+  stopifnot(is.character(search_terms))
 
-  # If length of loc is more than one, transform values into a regex friendly
-  # string.
-  if (length(loc) > 1) {
-    loc <- paste(loc, collapse = "|")
+  # If length of search_terms is more than one, transform values into a regex
+  # friendly string.
+  if (length(search_terms) > 1) {
+    search_terms <- paste(search_terms, collapse = "|")
   }
 
   # Fetch the country databse from ct_env.
@@ -38,7 +44,7 @@ ct_country_lookup <- function(loc, type = c("reporter", "partner")) {
 
   # Perform the lookup.
   ans <- country_df[country_df$type == type &
-                      grepl(loc, country_df$`country name`,
+                      grepl(search_terms, country_df$`country name`,
                             ignore.case = TRUE),
                     c("country name")]
 
