+++ tests/testthat/test-ct_country_lookup.R
@@ -2,7 +2,7 @@ context("ct_country_lookup")
 
 
 test_that("correct vals given reporter lookup", {
-  vals <- ct_country_lookup(loc = c("Korea", "EU"), type = "reporter")
+  vals <- ct_country_lookup(search_terms = c("Korea", "EU"), type = "reporter")
 
   # Check values of output.
   expect_equal(vals, c("Dem. People's Rep. of Korea",
@@ -12,7 +12,7 @@ test_that("correct vals given reporter lookup", {
 
 
 test_that("correct vals given partner lookup", {
-  vals <- ct_country_lookup(loc = c("Korea", "EU"), type = "partner")
+  vals <- ct_country_lookup(search_terms = c("Korea", "EU"), type = "partner")
 
   # Check values of output.
   expect_equal(vals, c("Africa CAMEU region, nes",
@@ -27,18 +27,18 @@ test_that("correct vals given partner lookup", {
 
 
 test_that("Correct vals given input not found in country DB", {
-  expect_equal(ct_country_lookup(loc = "not_a_country",
+  expect_equal(ct_country_lookup(search_terms = "not_a_country",
                                  type = "reporter"),
                "No matching results found")
 })
 
 
 test_that("errors and warnings are thrown as expected", {
-  # Throw error with invalid input for param "loc".
-  expect_error(ct_country_lookup(loc = 533,
+  # Throw error with invalid input for param "search_terms".
+  expect_error(ct_country_lookup(search_terms = 533,
                                  type = "reporter"))
 
   # Throw error with invalid input for param "type".
-  expect_error(ct_country_lookup(loc = c("Korea", "EU"),
+  expect_error(ct_country_lookup(search_terms = c("Korea", "EU"),
                                  type = "not_reporter"))
 })
