### 2017-09-21 518f089
bug fix with arg 'trade_direction'
