### 2017-09-21 fd4c8d1
bug fix in how commodity_df is saved to file
