### 2017-09-21 8dace54
add input validation
