### 2017-09-21 b921e7b
add new tests, split existing tests into smaller chunks
