+++ tests/testthat/test-ct_commodity_db_type.R
@@ -1,0 +1,5 @@
+context("ct_commodity_db_type")
+
+test_that("correct return value", {
+  expect_equal(ct_commodity_db_type(), "HS")
+})
