+++ tests/testthat/test-ct_get_reset_time.R
@@ -1,0 +1,5 @@
+context("ct_get_reset_time")
+
+test_that("same test, different iteration", {
+  expect_is(ct_get_reset_time(), "POSIXct")
+})
