+++ DESCRIPTION
@@ -12,8 +12,7 @@ Imports:
         jsonlite, 
         magrittr (>= 1.5), 
         methods, 
-        purrr, 
-        utils
+        purrr
 RoxygenNote: 6.0.1
 URL: https://github.com/ChrisMuir/comtradr
 BugReports: https://github.com/ChrisMuir/comtradr/issues
