### 2017-09-21 ae907f0
update test for better code coverage
