### 2017-09-21 57ccd92
expand test coverage
