### 2017-09-21 f9ce53e
update README to reflect function edits
