+++ README.md
@@ -3,7 +3,7 @@
 comtradr
 ========
 
-[![Travis-CI Build Status](https://travis-ci.org/ChrisMuir/comtradr.svg?branch=master)](https://travis-ci.org/ChrisMuir/comtradr) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ChrisMuir/comtradr?branch=master&svg=true)](https://ci.appveyor.com/project/ChrisMuir/comtradr) [![codecov](https://codecov.io/github/ChrisMuir/comtradr/branch/master/graphs/badge.svg)](https://codecov.io/github/ChrisMuir/comtradr) [![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/comtradr)](http://cran.r-project.org/package=comtradr)
+[![Travis-CI Build Status](https://travis-ci.org/ChrisMuir/comtradr.svg?branch=master)](https://travis-ci.org/ChrisMuir/comtradr) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ChrisMuir/comtradr?branch=master&svg=true)](https://ci.appveyor.com/project/ChrisMuir/comtradr) [![codecov](https://codecov.io/github/ChrisMuir/comtradr/branch/master/graphs/badge.svg)](https://codecov.io/github/ChrisMuir/comtradr) [![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/comtradr)](https://cran.r-project.org/package=comtradr)
 
 R package for interacting with the [UN Comtrade Database](https://comtrade.un.org/data/) public API. UN Comtrade provides historical data on the weights and value of specific goods shipped between countries, more info can be found [here](https://comtrade.un.org/). Full API documentation can be found [here](https://comtrade.un.org/data/doc/api/).
 
@@ -48,46 +48,49 @@ ct_country_lookup("korea")
 # Since we want South Korea, we'll use "Rep. of Korea" within the API query.
 example1 <- ct_search(reporters = "China", 
                       partners = c("Rep. of Korea", "USA", "Mexico"), 
-                      tradedirection = "exports")
+                      trade_direction = "exports")
 
 # Inspect the return data
-str(example1$data)
+str(example1)
 #> 'data.frame':    75 obs. of  35 variables:
-#>  $ Classification        : chr  "H4" "H4" "H4" "H4" ...
-#>  $ Year                  : int  2012 2012 2012 2013 2013 2013 2014 2014 2014 2015 ...
-#>  $ Period                : int  2012 2012 2012 2013 2013 2013 2014 2014 2014 2015 ...
-#>  $ Period Desc.          : chr  "2012" "2012" "2012" "2013" ...
-#>  $ Aggregate Level       : int  0 0 0 0 0 0 0 0 0 0 ...
-#>  $ Is Leaf Code          : int  0 0 0 0 0 0 0 0 0 0 ...
-#>  $ Trade Flow Code       : int  2 2 2 2 2 2 2 2 2 2 ...
-#>  $ Trade Flow            : chr  "Export" "Export" "Export" "Export" ...
-#>  $ Reporter Code         : int  156 156 156 156 156 156 156 156 156 156 ...
-#>  $ Reporter              : chr  "China" "China" "China" "China" ...
-#>  $ Reporter ISO          : chr  "CHN" "CHN" "CHN" "CHN" ...
[…]
