### 2017-09-21 9793227
add test for when hourly query limit is 0
