### 2017-09-22 12b5c16
Add more tests
