### 2017-09-22 f274158
remove unused arg from examples
