+++ R/ct_commodity_lookup.R
@@ -37,14 +37,12 @@
 #' @examples
 #' # Look up commodity descriptions related to "halibut"
 #' ct_commodity_lookup("halibut",
-#'                     commoditydf,
 #'                     return_code = FALSE,
 #'                     return_char = FALSE,
 #'                     verbose = TRUE)
 #'
 #' # Look up commodity codes related to "shrimp".
 #' ct_commodity_lookup("shrimp",
-#'                     commoditydf,
 #'                     return_code = TRUE,
 #'                     return_char = FALSE,
 #'                     verbose = TRUE)
