### 2017-09-22 703a519
remove 'dontrun' tags from examples
