### 2017-09-23 81eb45d
typo fix in the examples
