### 2017-09-23 fcacb30
use attributes with the reference table datasets
