### 2017-09-23 a68ca35
rework how the col_names are handled
