+++ R/ct_search.R
@@ -35,11 +35,10 @@
 #' @param url Base of the Comtrade url string, as a character string. Default
 #'  value is "https://comtrade.un.org/api/get?" and should mot be changed
 #'  unless Comtrade changes their endpoint url.
-#' @param col_name Should the output data frame have human-friendly or
-#'  machine-friendly column names. Human-friendly means easy for a human to
-#'  read and understand. Machine-friendly means easy for a machine to parse,
-#'  and may not be intuitive for a human to read and understand. Must be
-#'  either "human" or "machine". Default value is "human".
+#' @param col_name char string, the type of the column headers to use. "desc"
+#'  will return headers that are descriptive and easy to interpret. "comtrade"
+#'  will return the col headers that are used by the UN Comtrade API. Both
+#'  options are machine-readable. Default value is "desc".
 #'
 #' @details Basic rate limit restrictions. For details on how to register a
 #'  valid token, see \code{\link{ct_register_token}}. For API docs on rate
@@ -118,7 +117,7 @@ ct_search <- function(reporters, partners,
                       commod_codes = "TOTAL", max_rec = NULL,
                       type = c("goods", "services"),
                       url = "https://comtrade.un.org/api/get?",
-                      col_name = c("human", "machine")) {
+                      col_name = c("desc", "comtrade")) {
 
   ## Input validation related to API limits on parameter combinations.
 
@@ -166,8 +165,7 @@ ct_search <- function(reporters, partners,
     stop(msg)
   }
 
-
-  # Transformations to type.
+  ## Transformations to type.
   type <- match.arg(type)
   if (type == "goods") {
     type <- "C"
@@ -175,8 +173,7 @@ ct_search <- function(reporters, partners,
     type <- "S"
   }
 
-
-  # Transformations to freq.
+  ## Transformations to freq.
   freq <- match.arg(freq)
   if (freq == "annual") {
     freq <- "A"
@@ -184,8 +181,7 @@ ct_search <- function(reporters, partners,
     freq <- "M"
   }
 
-
-  # Transformations to start_date and end_date.
+  ## Transformations to start_date and end_date.
   if (any(c(start_date, end_date) %in% c("all", "All", "ALL"))) {
     date_range <- "all"
   } else {
@@ -224,8 +220,7 @@ ct_search <- function(reporters, partners,
     date_range <- paste(date_range, collapse = ",")
   }
 
-
-  # Transformations to reporters.
+  ## Transformations to reporters.
   if (any(reporters %in% c("all", "All", "ALL"))) {
     reporters <- "All"
   }
@@ -250,8 +245,7 @@ ct_search <- function(reporters, partners,
   }) %>%
     paste(collapse = ",")
 
-
-  # Transformations to partners.
+  ## Transformations to partners.
   if (any(partners %in% c("all", "All", "ALL"))) {
     partners <- "All"
   }
@@ -276,8 +270,7 @@ ct_search <- function(reporters, partners,
   }) %>%
     paste(collapse = ",")
 
-
-  # Transformations to trade_direction.
+  ## Transformations to trade_direction.
   if (any(tolower(trade_direction) == "all")) {
     trade_direction <- "all"
   } else {
@@ -315,8 +308,7 @@ ct_search <- function(reporters, partners,
     trade_direction <- rg
   }
 
-
-  # Transformations to commod_codes.
+  ## Transformations to commod_codes.
   stopifnot(is.character(commod_codes))
   if (any(tolower(commod_codes) == "total")) {
     commod_codes <- "TOTAL"
@@ -329,23 +321,12 @@ ct_search <- function(reporters, partners,
     commod_codes <- paste(commod_codes, collapse = ",")
   }
 
-
-  # Transformations to colname:
[…]
