### 2017-09-23 #141 @ChrisMuir
@AliciaSchep and @rtaph,

So just to update, I've been working on implementing changes and edits to `comtradr` based on your reviews. I wanted to outline the changes made so far.

Before getting into it, I just wanted to say I've learned A LOT during this review and while working on refining the package. Thank you both again for taking the time.

The tl:dr version of changes is [here](https://github.com/ChrisMuir/comtradr/issues/4).

# Overview of Major Changes
- Eliminated functions `ct_commodities_table`, `ct_countries_table`, `ct_json_data`, `ct_csv_data`.
- Added new exported functions `ct_commodity_db_type`, `ct_update_databases`, `ct_get_remaining_hourly_queries`, `ct_get_reset_time`, `ct_register_token`.
- A package environment (`ct_env`) is created upon load, that's used to handle variables related to API rate limits as well as package data.
- Saved the country table DB and commodity table DB as pkg data files to directory `inst/extdata`, both are read-in to `ct_env` upon pkg load, and accessed by package functions whenever needed.
- Created a framework for throttling API requests per the API limits. For now this only applies to function `ct_search`.
- Add function that gives users the ability to set a valid API key/token via global options.

# Replies to review comments

> I liked the example plots and manipulations in the README, although those might be more appropriate for the vignette rather than README.

This makes good sense, the two examples using  `ggplot2` have been moved to the vignette.

> Examples in package documentation did not run successfully using devtools::run_examples(run = FALSE). It looked like there was some R output included in the documentation -- I think that should be commented out if included. I am guessing the examples were set to not run for purposes of CRAN, but it would be nice to be able to run them locally in an automated way.

Yeah this was originally done for CRAN submission, but I've removed the `dontrun` tags from all functions that do not require an internet connection.

> The package should ideally have a man page for the entire package, which might describe briefly what the package does.

This has been added.

> The tests include many expectations each, especially for ct_search. This can make it harder to figure out exactly why the test is failing, because you get the message from the "test_that" call but then have to do some digging to figure out exactly which expectation(s) failed.

I split up as many of the tests as I could, so they're now more manageable.

> The first time I ran the tests, all the expectations for "ex2" failed. Trying again multiple times, it seemed to work. I am not sure what went wrong the first time, so don't really have any suggestion for that one, just figured I'd mention it.

@rtaph 's assessment was correct, the `ct_search` tests were failing so often because they were violating the per second limit of the API (I was really surprised by this!). The package now incorporates rate limiting, so the tests are much more stable.

> It would be good to include the Travis Badge on the README, and to additionally test the package on MacOs and Windows. For Mac, this can be specified in the .travis.yml file. For Windows, you can use AppVeyor.

I've expanded Travis CI to now build on Linux and Mac, and have added AppVeyor for Windows CI (plus all the badges!).

> Most of the package names start with "ct_", except the "country_lookup" and "commodity_lookup" functions. I think it would be nice to go fully with the convention of starting all the functions with the same prefix, and add "ct_" to those two functions as well.

This was a great idea, all exported functions now have the `ct_` prefix.

> For the country_lookup function, I think it would be nice for the order of the arguments to be re-arranged so that "type" was optional without having to write-out the "lookuptable" argument.

Function `ct_country_lookup` now no longer has arg `lookuptable`, but this suggestion was mentioned by both reviews, so I took care to consider the arg order for all pkg functions, and made edits to the order within `ct_search`.

> Unit tests exist, but a whole battery of ct_search tests fail when you run the full test suite. This is because the API limits the query rate in guest mode and the tests are all run quickly in succession. As a result, the tests all fail. When I run the tests one by one in the console, they pass. They also pass when I pad your test suite with Sys.sleep(1) in between calls to the API. More on this below.

This has been addressed.

> The package does not use the recommended snake_case convention and mixes naming styles (see below). The standard is not hard and fast but it would be nice if a single approach was adopted, perhaps working around the mixed naming conventions in the API.

This has been fixed, `snake_case` is used throughout the package.

> the package does not provide citation details in the README, as per the guideline

I had a question about this and adding a date to `DESCRIPTION`. Per [this](https://github.com/ropensci/onboarding/issues/72#issuecomment-258560452) comment from @sckott, adding date to `DESCRIPTION` is not necessary? Just wanted to get clarity on this.

> the package does not yet contain top-level documentation (i.e. typing ?comtrade will not return a help page)

This has been added.

> per guidelines, the package ought to add #' @noRd to internal functions such as api_col_names()

Done, I've added `#' @noRd` to all internal functions.

> Returning a Data Frame Directly

I thought this was great, I had always felt kinda weird about returning a list. I made this change.

> Converting to Native R Messages and Warnings

This change has been made. I'm checking the response from `httr::GET` to make sure the status code is 200, checking that the response type is json, and if no data was returned I'm checking to see if a helpful error message was provided by the API. Errors are thrown, as opposed to wrapping the info up as part of the return list object.

> Caching Reference Tables

This has been implemented. The package now features the country reference table and commodity reference table saved as .rda files to `inst/extdata`. They are both read-in when the pkg loads, and the functions utilize them whenever needed. The user is no longer required to download them to begin making API requests. Users can check for reference table updates with function `ct_update_databases`, and this function also gives the user the option of replacing the commodity table on file with any of the other types of commodity tables available ("HS", "HS1992", "SITC", etc).

> Throttling API requests

This is done. Upon package load, three variables are created in pkg env `ct_env`, they're basically "time of last query", "number of available queries left this hour", and "time that the current hour chunk resets". These variables allow the package to keep API requests under the limits of 1/second and 100/hour.

> I would avoid recyling the package name in example assignments (comtrade <-)

I fixed this.

> The documentation is generally good with plenty of examples. A few topics I think could be added to the documention include:
Oauth Token: documenting the use of an authorization token.
Regex: It would be great to state in the documentation that the functions use regular expressions. Also, not all your users may know regex, so pointing them to some help pages on the topic could be great (maybe add to the see-also section of documention).

I've added a section in the vignette on rate limits and using function `ct_register_token`.
For functions `ct_country_lookup` and `ct_commodity_lookup`, I give a brief explanation of the use of regex in the functions, provide a link to my favorite [regex tutorial](http://stat545.com/block022_regular-expression.html), and added `grepl` to the "seealso" section.

> The one thing I would split out is the part of the function that constructs the API call from the part that actually calls it. It would be nice to be able to generate the API url for troubleshooting purposes, even if this is not exported, and also save the API url along with the result (especially if caching).

I have the tasks of create the API url and actually make the request split into two separate functions, but they aren't as modular as what you're getting at here. I did add to `ct_search` that the API url is now attached to the return data frame as an attribute.

> It is good web etiquette to identify oneself when accessing APIs and scraping the web. This can be done via the user agent variable in httr and would be good for people using guest mode.

This is done, using the method that was suggested:
```r
paste(Sys.info()[["user"]], R.version$version.str, version$platform, sep = ", ")
```

> Is there any utility in allowing the user to specify fmt? I recommend removing the flexibility and only letting the API receive JSON

Nope there was no upside to using the "csv" option previously. I've removed it, and thus removed arg `fmt` from `ct_search`.

> From what I can tell, there is no call to check the API version (is that right?). If there is, it would be nice assert the version in the return data so that if the API changes, the user gets a warning about using an older version of comtradr.

Is this suggestion in regard to the API version or the version of the `comtradr` package? I don't believe the API itself features versioning.

> This is really a matter of preference, but I would simplify by removing the option of "human readable" names in the formals, instead providing a dictionary and examples of how to do this outside of the function. This lets the user go back and forth with renaming columns once the data frame is returned.
Another thought is that the "country name" header has a space, and some people may not know how to use backticks to refer to names like this. Again, maybe think about standardizing the naming convention for data frame headers, argument names, and argument parameters.

I edited all col names within `api_col_names` to be strictly `snake_case`. I'm not sure what you mean by providing a dictionary.....like provide a vector of names as a package data object? For now I've left the `api_col_names` function in, and left in the arg `col_name` within `ct_search`.

I also edited the column names approach slightly, the args are now `desc` and `comtrade`. "desc" will return headers that are descriptive and easy to interpret. "comtrade" will return the col headers that are used by the UN Comtrade API. Both options are machine-readable. Default value is "desc" within `ct_search`.

> Why not have an asserter to ensure that your receive Status: 200? Maybe this is not needed since you check the validation codes for JSON, but still, it is easy to implement via rawdata$status_code == 200L and an added layer of safety.

I'm checking for status code 200 in function `execute_api_request` (error is thrown if not 200).

> spelling of "ajust" in comtradr-vignette.Rmd, line 98.
spelling of "tomatos" in comtradr-vignette.Rmd, lines 90 and 94

These are fixed.

> Is there any reason that skip_on_cran() is commented out? This seems to be the sort of package that this function is useful for.

Yeah at some point after getting on CRAN I was having issues with the tests being skipped on Travis, and commenting out `skip_on_cran` seemed to fix the issue. I've added `skip_on_cran` back to the functions that require an internet connection.

> Why not save api_col_names as internal package data instead of a function? Also, why not save it as a named list (with machine readable variants) in order to allow the end user to have a dictionnary to go back and forth between those names?

Again, I'm not sure I understand. If I include the col names as package data, I assume I would also export a function for accessing and using it?

> It would be great if the following metadata was stored somewhere along the way, including:
the API timestamp: When was the data called?
the API URL: What is the address/definition of the call?
system.time() timings: How long did the operation take to complete?

This has been done. I've added these three items as attributes to the return data frame from `ct_search`.

> There is no date field in DESCRIPTION file. It is not strictly necessary, but it means that running citation("comtradr") will not return a proper citation.

See my earlier comment on adding date to `DESCRIPTION`.

> I suggest setting the order of your function's formals in order of importance (or frequency of use). For example, I don't think the api URL is a common call, so moving it down would help for those who like to use positional matching.

I reordered the function args in `ct_search` to address this.

> Why not let the verbose argument in your functions draw from the existing setting in the options? I.e. verbose = getOption("verbose", TRUE)

I wasn't sure about this, and have left it as is for now. I think I prefer the way it's set up now, but that's just me (if I knew that most users disagreed, I would certainly be inclined to change it). Is this typically handled using options in the major R packages?

> The country table is not denormalized (same code if both a reporter and partner). This is nitpicking but I think a nicer approach would be a single row for each country + code, with logical columns that define if they are a reporter and if they are a partner.

I initially misread this comment and am just now realizing what you're getting at, so I haven't done anything to address this. Just to be clear, and using Germany as an example, you're saying go from this:
```r
> country_df[country_df$`country name` == "Germany", ]
    code country name     type
101  276      Germany reporter
373  276      Germany  partner
```
to this:
```r
> country_df[country_df$`country name` == "Germany", ]
  code country name reporter partner
1  276      Germany     TRUE    TRUE
```

> Since I do not think you are using the dots ... argument for anything, why not add them in to allow users to pass additional parameters to the regex functions (for example, to modify the ignore.case or perl arguments)?

I haven't implemented this. I like the idea, but for both of the `lookup` functions, I would like for the default setting for `ignore.case` within `grepl` to be TRUE. I could add `ignore.case` to the input args for both functions along with adding the dots, so:
```r
ct_commodity_lookup(search_terms, return_code = FALSE, return_char = FALSE, verbose = TRUE, ignore.case = TRUE, ...)
```
That seems weird to me though, to say "`...` is for additional args that will be passed along to `grepl`, except for `ignore.case`, as that's defined explicitly". Is there a better way to go about this?

> Questions
Will fromJSON() with simplifyDataFrame = TRUE always work? Is there a way to test for non-tabular data, or catch errors? What if there is hierarchical nesting?

In all of the testing I've done, it's never been an issue. In the cases in which there is legitimately no data to return based on the input values (and thus no error will be thrown), within internal function `execute_api_request` an empty data frame is created and returned with the specified column names.

# Edits I'm planning to make
- Add section to vignette explaining function `ct_update_databases`.
