### 2017-09-23 724d723
add test for arg col_name in func ct_search
