### 2017-09-24 #141 @rtaph
Props @ChrisMuir on implementing so many major changes so quickly! I installed https://github.com/ChrisMuir/comtradr/commit/724d723dff33fae4384f3e10c7c43adfa985c30c and took your package for spin. Everything works nicely!

Based on your comments, I have some follow-on remarks:

 > Is this suggestion in regard to the API version or the version of the comtradr package? I don't believe the API itself features versioning.

Yes, my point was related to the API version (not the `comtradr` package version). I couldn't find any documentation from the original API page and thought you might know. It would have been nice to be able to query the API to get the API version but if the developers never included this, this isn't implementable.

> That seems weird to me though, to say "... is for additional args that will be passed along to grepl, except for ignore.case, as that's defined explicitly". Is there a better way to go about this?

Personally I like your choice of setting `ignore.case = TRUE` since I anticipate almost use cases to be case-insensitive. Your suggested language sounds fine to me. However, I would add roxygen documentation for the `ignore.case` argument, simply saying "argument passed to grepl()" and hyperlinking to the grepl man page.

>  I'm not sure what you mean by providing a dictionary.....like provide a vector of names as a package data object? For now I've left the api_col_names function in, and left in the arg col_name within ct_search.

Yes, I am thinking of python with my use of 'dictionary'-- namely, a list (or vector) of uniquely named key-value pairs. My point, really, was two-fold:

1. Saving the information as [package sysdata](http://r-pkgs.had.co.nz/data.html#data-sysdata) seemed more conventional than writing a static function to do the same thing. Furthemore, defining a named list/vector is visually more maintainable compared to creating two unnamed vectors that need to matched positionally. It makes the binary relation explicit.
2. Exposing the key-value pair(s) as [exported data](http://r-pkgs.had.co.nz/data.html#data-data) is useful to the analyst, who will likely need to make conversions between names.

To illustrate with a 2-entry example

```r
# API dictionary, for internal use since the user does not need to interact
#  with the the original API naming
h1 <- list(classification = "pfCode",
           trade_flow     = "rgDesc")
devtools::use_data(h1, internal = TRUE)

# Human-readable dictionary, exported
h2 <- list(`Classification` = "classification",
           `Trade Flow`     = "trade_flow")
h2 <- c(`Classification` = "classification", # or as a vector
        `Trade Flow`     = "trade_flow")
devtools::use_data(h2)
```

I guess I liked that you previously had human-readable variants of names, but thought the implementation would be stronger if  this mapping was exposed to the user to use as they saw fit. A dictionary would let users do something like `dplyr::rename_(ex_1, .dots = h2)` and use these 'polished' names for plotting, publication tables, etc. For me as a user, I would likely want to use both the `snake_case` and human-readable variants for different purposes. This should be achievable as a post-query operation.

Additionally, setting `col_name = "comtrade"` in `ct_search` returns the API naming convention, which I think is much less useful than the human readable names you had previously (Hence my suggestion of having two dictionaries).

> Will fromJSON() with simplifyDataFrame = TRUE always work?  ... In all of the testing I've done, it's never been an issue

Okay great! I just have worked with (scraped) HTML and XML data in the past where dataframe simplification arguments failed. You know the API well so if it's not a problem you have encountered, then it's a non-issue!

> verbose = getOption("verbose", TRUE) ... I wasn't sure about this, and have left it as is for now.

Ok!

> I initially misread this comment and am just now realizing what you're getting at, so I haven't done anything to address this. Just to be clear, and using Germany as an example, you're saying go from this ... to this

Yes, exactly. It's a small change, but I think makes it easier for a new user to grasp the data structure and filter with logicals.

> the package does not provide citation details in the README, as per the guideline

Interesting about the [date field](https://github.com/ropensci/onboarding/issues/72#issuecomment-258560452)... I did not know about the CRAN behaviour! I defer to @sckott on adding the date field to the `DESCRIPTION`, file. But if I understand correctly, this decision might be separate from adding citation guidance in the README.

A few last suggestions:
- I recommend changing `2nd_partner` to `second_partner` since valid R names cannot start with a digit. Same with `2nd_partner_code` and `2nd_partner_iso`.
- A number of fields return empty in JSON data which I believe would be better represented by `NA` instead of empty character strings. What do you think?
-  What is the purpose of the `period_desc` variable from the API? Since it is identical to `period`, can it be removed? I ask because it's redundant, and I don't like the idea of a column name describing the ordering of the data (since this can change).
