### 2017-09-28 881607d
add ignore.case as func arg
