+++ tests/testthat/test-ct_commodity_lookup.R
@@ -62,11 +62,19 @@ test_that("correct vals given return_code == FALSE & return_char == FALSE", {
   expect_named(vals, c("tomato", "trout"))
 })
 
+
+test_that("throw warning when ignore.case = TRUE and fixed = TRUE", {
+  expect_warning(ct_commodity_lookup(search_terms = "trout",
+                                     ignore.case = TRUE, fixed = TRUE))
+})
+
+
 test_that("throw warning with invalid input to arg 'search_terms'", {
   expect_warning(ct_commodity_lookup(search_terms = c("not_a_trade_item",
                                                       "also_not_a_trade_item"),
                                      return_char = TRUE))
 })
+
 
 test_that("throw error with invalid input to arg 'value'", {
   expect_error(ct_commodity_lookup(search_terms = list()))
