### 2017-09-28 af98f3f
change structure of country table, add ignore.case as func arg
