+++ R/ct_update_databases.R
@@ -187,92 +187,61 @@ ct_update_databases <- function(force = FALSE, verbose = TRUE,
     }
   }
 
-  # Get the reporter country database from the Comtrade website. Compare the
-  # "last-modified" date value within the header to the value in variable
-  # "date" for reporter countries within the current country databse. If the
-  # "last-modified" date is newer than the date within the current country
-  # DB, the reporter countries portion of the old DB will be replaced by
-  # the newer DB, both for the current session and within the data dir of
-  # the comtradr package.
+  # Get the reporter country database and the partner country database from
+  # the Comtrade website. Compare the "last-modified" date value within the
+  # header of each to the "date" attribute of the current country_table
+  # reference dataset on file. If either "last-modified" date is newer than
+  # the date attr of the dataset on file, replace the dataset on file with the
+  # data pulled from the Comtrade website.
   country_update <- FALSE
-  res <- httr::GET(reporter_url, httr::user_agent(get("ua", envir = ct_env)))
+  res_rep <- httr::GET(reporter_url,
+                       httr::user_agent(get("ua", envir = ct_env)))
+  res_par <- httr::GET(partner_url,
+                       httr::user_agent(get("ua", envir = ct_env)))
   if (force ||
-      httr::headers(res)$`last-modified` >
-      attributes(country_df)$reporter_date) {
-    # Extract data frame.
-    df <- res %>%
+      httr::headers(res_rep)$`last-modified` > attributes(country_df)$date ||
+      httr::headers(res_rep)$`last-modified` > attributes(country_df)$date) {
+    country_update <- TRUE
+    # Get reporters dataset as data frame.
+    reporters <- res_rep %>%
       httr::content("text", encoding = "UTF-8") %>%
       jsonlite::fromJSON(simplifyDataFrame = TRUE) %>%
       magrittr::extract2("results") %>%
-      `colnames<-`(c("code", "country name"))
-    df$type <- "reporter"
-    # Replace the reporters portion of the current country_df with the new
-    # data.
-    if (!force) {
-      country_update <- TRUE
-      country_df[country_df$type == "reporter", ] <- df
-      attributes(country_df)$reporter_date <- curr_date
-    } else {
-      country_df <- df
-      attributes(country_df)$reporter_date <- curr_date
-    }
-    # Update the output message.
-    if (verbose) {
-      if (grepl("Updates found", msg, fixed = TRUE)) {
-        msg <- paste0(msg, ", reporter countries")
-      } else {
-        msg <- paste0("Updates found. The following datasets have been ",
-                      "downloaded: reporter countries")
-      }
-    }
-  }
-
-  # Get the partner country database from the Comtrade website. Compare the
-  # "last-modified" date value within the header to the value in variable
-  # "date" for partner countries within the current country database. If the
-  # "last-modified" date is newer than the date within the current country
-  # DB, the partner countries portion of the old DB will be replaced by
-  # the newer DB, both for the current session and within the data dir of
-  # the comtradr package.
-  res <- httr::GET(partner_url, httr::user_agent(get("ua", envir = ct_env)))
-  if (force ||
-      httr::headers(res)$`last-modified` >
-      attributes(country_df)$partner_date) {
-    # Extract data frame.
-    df <- res %>%
+      `colnames<-`(c("code", "country_name"))
+    # Get partners dataset as data frame.
+    partners <- res_par %>%
       httr::content("text", encoding = "UTF-8") %>%
       jsonlite::fromJSON(simplifyDataFrame = TRUE) %>%
       magrittr::extract2("results") %>%
-      `colnames<-`(c("code", "country name"))
-    df$type <- "partner"
-    ###########df$date <- curr_date
[…]
