### 2017-09-29 463d640
replace '2nd' with 'second' in col names
