+++ R/utils.R
@@ -25,9 +25,9 @@ api_col_names <- function(type = c("desc", "comtrade")) {
       "partner_code",
       "partner",
       "partner_iso",
-      "2nd_partner_code",
-      "2nd_partner",
-      "2nd_partner_iso",
+      "second_partner_code",
+      "second_partner",
+      "second_partner_iso",
       "customs_proc_code",
       "customs",
       "mode_of_transport_code",
