### 2017-10-02 69f6883
in returned data, replace all empty str with NA
