### 2017-10-03 0d74c6f
output snake_case col headers, add ct_pretty_cols
