+++ R/comtradr.R
@@ -35,3 +35,18 @@
 #' @docType package
 #' @name comtradr
 NULL
+
+#' "pretty" column headers for Comtrade API data.
+#'
+#' Named vector of polished column headers, intended for use with plots,
+#' publication tables, etc.
+#'
+#' @docType data
+#' @name ct_pretty_cols
+#' @format Named vector, with the polished column headers as the names, and
+#'  the machine-readable column headers as the values. Each element is meant
+#'  to be treated as a key-value pair. The function \code{\link{ct_search}}
+#'  returns data with the machine-readable column headers by default.
+#' @examples
+#' data(ct_pretty_cols)
+NULL
