+++ man/ct_use_pretty_cols.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ct_use_pretty_cols.R
+\name{ct_use_pretty_cols}
+\alias{ct_use_pretty_cols}
+\title{Use Pretty Column Headers}
+\usage{
+ct_use_pretty_cols(df)
+}
+\arguments{
+\item{df}{data frame, Comtrade API data frame, returned from function
+\code{\link{ct_search}}.}
+}
+\value{
+data frame, input df with polish column headers.
+}
+\description{
+Transform the column headers of return data from funcion
+\code{\link{ct_search}} into a more "polished" set of column headers.
+Intended for use with plots, publication tables, etc.
+}
+\examples{
+\dontrun{
+# Pull API data
+df <- ct_search("Germany", "Canada")
+
+# Use polished column names
+df <- ct_use_pretty_cols(df)
+}
+}
