+++ NAMESPACE
@@ -8,4 +8,5 @@ export(ct_get_reset_time)
 export(ct_register_token)
 export(ct_search)
 export(ct_update_databases)
+export(ct_use_pretty_cols)
 importFrom(magrittr,"%>%")

[…]
