### 2017-10-04 f7ff001
add test for ct_use_pretty_cols
