### 2017-10-05 ba9b63d
update docs to reflect new functions
