+++ vignettes/comtradr-vignette.Rmd
@@ -62,7 +62,7 @@ q <- ct_search(reporters = "USA",
                freq = "monthly")
 ```
 
-Countries passed to parameters `reporters` and `partners` must be spelled as they appear in the country code lookup table. Function `ct_country_lookup` allows us to query the country code look up table.
+Countries passed to parameters `reporters` and `partners` must be spelled as they appear in the Comtrade country reference table. Function `ct_country_lookup` allows us to query the country reference table.
 ```{r}
 ct_country_lookup("korea", "reporter")
 ct_country_lookup("bolivia", "partner")
@@ -73,11 +73,11 @@ q <- ct_search(reporters = "Rep. of Korea",
                trade_direction = "all")
 ```
 
-Search trade related to specific commodities (say, tomatoes). We can query the commodity database to see all of the different commodity descriptions available for tomatoes.
+Search trade related to specific commodities (say, tomatoes). We can query the Comtrade commodity reference table to see all of the different commodity descriptions available for tomatoes.
 ```{r}
 ct_commodity_lookup("tomato")
 ```
-If we want to search for shipment data on all of the commodity descriptions listed, then we can simply adjust the params for `ct_commodity_lookup` so that it will return only the codes, which can then be passed along to `ct_search`.
+If we want to search for shipment data on all of the commodity descriptions listed, then we can simply adjust the parameters for `ct_commodity_lookup` so that it will return only the codes, which can then be passed along to `ct_search`.
 ```{r, eval = FALSE}
 tomato_codes <- ct_commodity_lookup("tomato", 
                                     return_code = TRUE, 
@@ -88,7 +88,7 @@ q <- ct_search(reporters = "USA",
                trade_direction = "all", 
                commod_codes = tomato_codes)
 ```
-On the other hand, if we wanted to exclude the juice and sauces from our search, we can pass a vector of the relevant codes to the API call.
+On the other hand, if we wanted to exclude juices and sauces from our search, we can pass a vector of the relevant codes to the API call.
 ```{r}
 q <- ct_search(reporters = "USA", 
                partners = c("Germany", "France", "Japan", "Mexico"), 
@@ -125,7 +125,7 @@ ct_commodity_lookup(c("tomato", "trout"), return_char = FALSE)
 ```
 
 
-For `ct_commodity_lookup`, if any of the input search terms return zero results, and parameter `verbose` is set to `TRUE`, a warning will be printed to console (set `verbose` to `FALSE` to turn off this feature).
+For `ct_commodity_lookup`, if any of the input search terms return zero results and parameter `verbose` is set to `TRUE`, a warning will be printed to console (set `verbose` to `FALSE` to turn off this feature).
 ```{r}
 ct_commodity_lookup(c("tomato", "sldfkjkfdsklsd"), verbose = TRUE)
 ```
@@ -134,7 +134,7 @@ ct_commodity_lookup(c("tomato", "sldfkjkfdsklsd"), verbose = TRUE)
 
 The Comtrade API imposes rate limits on both guest users and premium users. `comtradr` features automated throttling of API calls to ensure the user stays within the limits defined by Comtrade. Below is a breakdown of those limits, API docs on these details can be found [here](https://comtrade.un.org/data/doc/api/#Authentication).
 
-* Without authentication token: 1 request per second, 100 requests per hour.
+* Without user token: 1 request per second, 100 requests per hour.
 * With valid user token: 1 request per second, 10,000 requests per hour.
 
 In addition to these rate limits, the API imposes some limits on parameter combinations.
@@ -149,8 +149,45 @@ Additionally, the maximum number of returned records from a single query without
 
[…]
