### 2017-10-05 e40c888
add info on how to cite pkg
