+++ README.md
@@ -15,7 +15,7 @@ Please [report](https://github.com/ChrisMuir/comtradr/issues) issues, comments, 
 
 Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 
-**NOTE**: As of 2017-06-27, the curl errors due to an inadility to validate the SSL certificate of the API site are no longer appearing, I believe that UN Comtrade fixed their SSL cert issue. As of 2017-07-03, the package is back on CRAN.
+For information on citation of this package, use `citation("comtradr")`
 
 Installation
 ------------
@@ -67,20 +67,20 @@ str(example1)
 #>  $ partner_code          : int  410 484 842 410 484 842 410 484 842 410 ...
 #>  $ partner               : chr  "Rep. of Korea" "Mexico" "USA" "Rep. of Korea" ...
 #>  $ partner_iso           : chr  "KOR" "MEX" "USA" "KOR" ...
-#>  $ 2nd_partner_code      : logi  NA NA NA NA NA NA ...
-#>  $ 2nd_partner           : chr  "" "" "" "" ...
-#>  $ 2nd_partner_iso       : chr  "" "" "" "" ...
-#>  $ customs_proc_code     : chr  "" "" "" "" ...
-#>  $ customs               : chr  "" "" "" "" ...
-#>  $ mode_of_transport_code: chr  "" "" "" "" ...
-#>  $ mode_of_transport     : chr  "" "" "" "" ...
+#>  $ second_partner_code   : logi  NA NA NA NA NA NA ...
+#>  $ second_partner        : chr  NA NA NA NA ...
+#>  $ second_partner_iso    : chr  NA NA NA NA ...
+#>  $ customs_proc_code     : chr  NA NA NA NA ...
+#>  $ customs               : chr  NA NA NA NA ...
+#>  $ mode_of_transport_code: chr  NA NA NA NA ...
+#>  $ mode_of_transport     : chr  NA NA NA NA ...
 #>  $ commodity_code        : chr  "TOTAL" "TOTAL" "TOTAL" "TOTAL" ...
 #>  $ commodity             : chr  "All Commodities" "All Commodities" "All Commodities" "All Commodities" ...
 #>  $ qty_unit_code         : int  1 1 1 1 1 1 1 1 1 1 ...
 #>  $ qty_unit              : chr  "No Quantity" "No Quantity" "No Quantity" "No Quantity" ...
+#>  $ alt_qty_unit_code     : logi  NA NA NA NA NA NA ...
+#>  $ alt_qty_unit          : chr  NA NA NA NA ...
 #>  $ qty                   : logi  NA NA NA NA NA NA ...
-#>  $ alt_qty_unit_code     : chr  "" "" "" "" ...
-#>  $ alt_qty_unit          : logi  NA NA NA NA NA NA ...
 #>  $ alt_qty               : logi  NA NA NA NA NA NA ...
 #>  $ netweight_kg          : logi  NA NA NA NA NA NA ...
 #>  $ gross_weight_kg       : logi  NA NA NA NA NA NA ...
@@ -89,8 +89,8 @@ str(example1)
 #>  $ fob_trade_value_usd   : logi  NA NA NA NA NA NA ...
 #>  $ flag                  : int  0 0 0 0 0 0 0 0 0 0 ...
 #>  - attr(*, "url")= chr "https://comtrade.un.org/api/get?max=50000&type=C&freq=A&px=HS&ps=all&r=156&p=410,842,484&rg=2&cc=TOTAL&fmt=json&head=H"
[…]
