### 2017-10-05 #141 @rtaph
Let me start by saying great work, @ChrisMuir ! It's easy to for us reviewers to pick apart a package, much harder to implement changes. I am recommending `comtradr` for approval. 🎉

If @sckott has any citation guidance, perhaps your README should change, but to me your proposed approach sounds reasonable.

Last suggestion (I hope): Removing ['inst/doc' from your `.gitignore`](https://github.com/STAT545-UBC/Discussion/issues/66) so that your vignettes are included with your package. At the moment they don't install via `devtools::install_github()` unless the user explicitly sets `build_vignettes = TRUE`. You've got such a nice vignette and you want people to find it with `browseVignettes("comtradr")`! It would also be great to link to the vignette in your `comtradr` man page, as that may be the starting point for users.
