### 2017-10-06 7a92402
tweak partners list in one of the api calls
