### 2017-10-06 294a09b
move vignette output files to inst/doc
