### 2017-10-06 cc5732d
add link to vignette in pkg man page
