+++ R/comtradr.R
@@ -5,6 +5,12 @@
 #'   of commodities, these functions allow for easy API query and data returned
 #'   as a tidy data frame.
 #'
+#' @section Package Vignette:
+#'
+#' \itemize{
+#'   \item \url{../doc/comtradr-vignette.html}
+#' }
+#'
 #' @section Documentation for the Comtrade API:
 #'
 #' \itemize{
@@ -30,6 +36,7 @@
 #'   \item \code{\link{ct_register_token}}
 #'   \item \code{\link{ct_search}}
 #'   \item \code{\link{ct_update_databases}}
+#'   \item \code{\link{ct_use_pretty_cols}}
 #' }
 #'
 #' @docType package
