### 2017-10-06 df09515
remove link to vignette from pkg man page
