+++ inst/doc/comtradr-vignette.Rmd
@@ -1,271 +1,0 @@
----
-title: "Intro to comtradr"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Intro to comtradr}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, echo = FALSE}
-knitr::opts_chunk$set(comment = "#>", collapse = TRUE, fig.width = 9, fig.height = 6)
-```
-
-## Package information
-
-API wrapper for the [UN Comtrade Database](https://comtrade.un.org/data/), which features inter-country trade data dating back to the early 1990's. Full API documentation can be found [here](https://comtrade.un.org/data/doc/api/). This package allows users to interact with the API directly from R, and features functions for making queries and importing data.
-
-## Install and load comtradr
-
-Install from CRAN:
-```{r eval = FALSE}
-install.packages("comtradr")
-```
-Or install the development version from GitHub:
-```{r eval = FALSE}
-devtools::install_github("ChrisMuir/comtradr")
-```
-Load comtradr
-```{r}
-library(comtradr)
-```
-
-## Making API calls
-Lets say we want to get data on all imports into the United States from Germany, France, Japan, and Mexico, for all years.
-```{r}
-q <- ct_search(reporters = "USA", 
-               partners = c("Germany", "France", "Japan", "Mexico"), 
-               trade_direction = "imports")
-
-# API calls return a tidy data frame..
-str(q)
-```
-
-Here are a few more examples to show the different parameter options:
-
-Limit the search range to shipments between 2010 and 2014.
-```{r, eval = FALSE}
-q <- ct_search(reporters = "USA", 
-               partners = c("Germany", "France", "Japan", "Mexico"), 
-               trade_direction = "imports", 
-               start_date = "2010-01-01", 
-               end_date = "2014-01-01")
-```
-
-By default, the return data is in yearly amounts. We can specify monthly, however the API limits each "monthly" query to a five month span.
-```{r, eval = FALSE}
-q <- ct_search(reporters = "USA", 
-               partners = c("Germany", "France", "Japan", "Mexico"), 
-               trade_direction = "imports", 
-               start_date = "2012-03-01", 
-               end_date = "2012-07-01", 
-               freq = "monthly")
-```
-
-Countries passed to parameters `reporters` and `partners` must be spelled as they appear in the Comtrade country reference table. Function `ct_country_lookup` allows us to query the country reference table.
-```{r}
-ct_country_lookup("korea", "reporter")
-ct_country_lookup("bolivia", "partner")
-```
-```{r, eval = FALSE}
-q <- ct_search(reporters = "Rep. of Korea", 
-               partners = "Bolivia (Plurinational State of)", 
-               trade_direction = "all")
-```
-
-Search trade related to specific commodities (say, tomatoes). We can query the Comtrade commodity reference table to see all of the different commodity descriptions available for tomatoes.
-```{r}
-ct_commodity_lookup("tomato")
-```
-If we want to search for shipment data on all of the commodity descriptions listed, then we can simply adjust the parameters for `ct_commodity_lookup` so that it will return only the codes, which can then be passed along to `ct_search`.
-```{r, eval = FALSE}
-tomato_codes <- ct_commodity_lookup("tomato", 
-                                    return_code = TRUE, 
-                                    return_char = TRUE)
-
-q <- ct_search(reporters = "USA", 
-               partners = c("Germany", "France", "Mexico"), 
-               trade_direction = "all", 
-               commod_codes = tomato_codes)
-```
-On the other hand, if we wanted to exclude juices and sauces from our search, we can pass a vector of the relevant codes to the API call.
-```{r}
-q <- ct_search(reporters = "USA", 
[…]
