+++ man/comtradr.Rd
@@ -11,14 +11,6 @@ Interface with and extract data from the United Nations
   of commodities, these functions allow for easy API query and data returned
   as a tidy data frame.
 }
-\section{Package Vignette}{
-
-
-\itemize{
-  \item \url{../doc/comtradr-vignette.html}
-}
-}
-
 \section{Documentation for the Comtrade API}{
 
 
