### 2017-10-06 #141 @ChrisMuir
So I've been working on the two vignette changes, but I'm having trouble getting the the HTML vignette included with the package. I've tried a few different things, using [this](https://github.com/STAT545-UBC/Discussion/issues/66) page and [this](http://stat545.com/packages03_activity_part2.html#vignette) page as references, but I keep running into the "rdb is corrupt" issue described [here](https://stackoverflow.com/questions/30424608/error-in-fetchkey-lazy-load-database) and discussed [here](https://github.com/hadley/devtools/pull/1001).

I'll keep working at it.
