### 2017-10-06 f9c9c20
update to reflect all rOpenSci changes, issue #4
