### 2017-10-06 #141 @ChrisMuir
I've updated the `NEWS` file to reflect all of the package changes. I'm close to a point in which I'm ready to submit the updated package to CRAN.... @sckott should I wait on that until this review process is finished?
