### 2017-10-11 #141 @sckott
for the vignette issue: Once it's on CRAN, vignette should be installed by default, so it won't be an issue unless users are installing dev version.

linking to vignette from a manual page is a constant problem, but can put a link to the online (github/CRAN) version on the man page 😄

Another option is to use pkgdown, but that's in addition to the vignette.
