### 2017-10-11 #141 @ChrisMuir
Hi Scott,

Thanks for the input. The package is currently on CRAN, this would just be submitting the updated version with all of the pkg review updates. I'll hold off on submission though, as the review is (I believe) close to being done.

I worked more on the vignette issue tonight, I think it's now fixed and all is working properly....vignette is now available after running `devtools::install_github`, and the package man page now links to the vignette.html file. I think the only issue was needing to run `devtools::build_vignettes` prior to pushing the edits to GitHub.

So I think all package edits have either been implemented or addressed 😄
