### 2017-10-18 #141 @sckott
Okay, on CRAN already, got it.

Great, will have one last quick look
