### 2017-10-19 34b0fa5
add rOpenSci footer
