### 2017-10-19 cc69a84
update links to reflect transfer of repo to rOpenSci
