### 2017-10-19 #141 @ChrisMuir
Great to hear! Thanks @sckott .

I've transferred the package to `ropensci`, and added the footer banner to the `README`, and updated all links in the `README` to point to the ropensci location of the repo.

On the blog post, I'd be interested, but my wife and I are moving to Nashville, TN in two weeks, and will be super busy for the next 3-4 weeks. So it would have to wait until after the move.

One question, should I add the "peer reviewed" badge to the repo?
