### 2017-10-19 #141 @ChrisMuir
Also, just want to say again to @rtaph and @AliciaSchep, I truly appreciate the help and the time that was put in. I think I learned more about building packages in the last month than I had learned in the previous two years 😄 , and the `comtradr` package is much better off for it.

Thanks so much!!
