### 2017-10-19 #141 @sckott
Great, thanks @ChrisMuir

We'll get in touch with you after the next month or so about the blog post.

yes! do add the peer review badge. like
