### 2018-03-22 #141 @ChrisMuir
Hi @AliciaSchep and @rtaph, hope you're both doing well! I just saw [this](https://ropensci.org/blog/2018/03/16/thanking-reviewers-in-metadata/) recent rOpenSci blog post... I've been thinking that `comtradr` is due for a release, and I would like to add you both to the `DESCRIPTION` file as reviewers prior to sending to CRAN.

So I wanted to find out if you're both okay with this. If so, do either of you have an [orcid.org](orcid.org) ID that you'd like me to include (like in [this](https://github.com/ropensci/codemetar/blob/master/DESCRIPTION) example).
