### 2018-03-22 #141 @ChrisMuir
Hi @AliciaSchep and @rtaph , hope you're both doing well! I just saw [this](https://ropensci.org/blog/2018/03/16/thanking-reviewers-in-metadata/) rOpenSci blog post about adding reviewers to a pkg.....I've been thinking `comtradr` is due for a new release, and I'd like to add you both as reviewers to the `DESCRIPTION` file prior to sending to CRAN.

So I wanted to see if this was okay with you both. If so, let me know if you also have an [orcid.org]() ID that you'd like me to include (like in [this](https://github.com/ropensci/codemetar/blob/master/DESCRIPTION) example).

Thanks!
