### 2018-03-22 #141 @ChrisMuir
Hi @rtaph , for your name in the `DESCRIPTION` file, should I use `person("Rafael", "P.H.", role = "rev", etc...)`, or would you prefer I use your full last name?
